Text summary

Create a reportable textual output for your model.

model_text(model, ...)

# S3 method for lm
model_text(model, ...)

Arguments

model
A statistical model.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

Examples

model <- t.test(Sepal.Length ~ Species, data = iris[1:100, ])
model_text(model)

#> The Welch Two Sample t-test suggests that the difference of Sepal.Length by Species (mean in group setosa = 5.01, mean in group versicolor = 5.94) is significant (difference = -0.93, 95% CI [-1.11, -0.75], t(86.54) = -10.52, p < .001) and can be considered as very large (Cohen's d = -2.26).

summary(model_text(model))

#> The Welch Two Sample t-test suggests that the difference of Sepal.Length by Species is significant (difference = -0.93, 95% CI [-1.11, -0.75], p < .001, Cohen's d = -2.26).

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
model_text(model)

#> We fitted a linear model (estimated using OLS) to predict Sepal.Length with Petal.Length and Species (formula = Sepal.Length ~ Petal.Length * Species). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Funder's (2019) recommendations.
#> 
#> The model explains a significant and substantial proportion of variance (R2 = 0.84, F(5, 144) = 151.71, p < .001, adj. R2 = 0.83). The model's intercept, corresponding to Sepal.Length = 0, Petal.Length = 0 and Species = setosa, is at 4.21 (SE = 0.41, 95% CI [3.41, 5.02], p < .001). Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and not significant (beta = 1.16, SE = 0.59, 95% CI [-0.01, 2.32], std. beta = 1.16, p = 0.052).
#>   - The effect of Speciesversicolor is negative and can be considered as medium and not significant (beta = -0.88, SE = 0.77, 95% CI [-2.41, 0.65], std. beta = 0.61, p = 0.256).
#>   - The effect of Speciesvirginica is negative and can be considered as very large and significant (beta = -1.75, SE = 0.79, 95% CI [-3.32, -0.18], std. beta = 0.97, p < .05).
#>   - The effect of Petal.Length:Speciesversicolor is positive and can be considered as very large and not significant (beta = 0.61, SE = 0.63, 95% CI [-0.63, 1.85], std. beta = -0.88, p = 0.334).
#>   - The effect of Petal.Length:Speciesvirginica is positive and can be considered as very large and not significant (beta = 0.97, SE = 0.62, 95% CI [-0.26, 2.19], std. beta = -1.75, p = 0.120).

summary(model_text(model))

#> We fitted a linear model to predict Sepal.Length with Petal.Length and Species.The model's explanatory power is substantial (R2 = 0.84, adj. R2 = 0.83). The model's intercept is at 4.21. Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and not significant (beta = 1.16, 95% CI [-0.01, 2.32], std. beta = 1.16, p = 0.052).
#>   - The effect of Speciesversicolor is negative and can be considered as medium and not significant (beta = -0.88, 95% CI [-2.41, 0.65], std. beta = 0.61, p = 0.256).
#>   - The effect of Speciesvirginica is negative and can be considered as very large and significant (beta = -1.75, 95% CI [-3.32, -0.18], std. beta = 0.97, p < .05).
#>   - The effect of Petal.Length:Speciesversicolor is positive and can be considered as very large and not significant (beta = 0.61, 95% CI [-0.63, 1.85], std. beta = -0.88, p = 0.334).
#>   - The effect of Petal.Length:Speciesvirginica is positive and can be considered as very large and not significant (beta = 0.97, 95% CI [-0.26, 2.19], std. beta = -1.75, p = 0.120).