Automatic reporting of R objects

Create reports of different objects. See the documentation for your object's class:

Usage

report(x, ...)

Arguments

x

The R object that you want to report (see list of of supported objects above).

...

Arguments passed to or from other methods.

Value

A list-object of class report, which contains further list-objects with a short and long description of the model summary, as well as a short and long table of parameters and fit indices.

Details

System and packages (sessionInfo)

Dataframes and vectors

Correlations and t-tests (htest)

ANOVAs (aov, anova, aovlist, ...)

Regression models (glm, lm, ...)

Mixed models (glmer, lmer, glmmTMB, ...)

Bayesian models (stanreg, brms...)

Bayes factors (from bayestestR)

Structural Equation Models (SEM) (from lavaan)

Model comparison (from performance())

Most of the time, the object created by the report() function can be further transformed, for instance summarized (using summary()), or converted to a table (using as.data.frame()).

Organization

report_table and report_text are the two distal representations of a report, and are the two provided in report(). However, intermediate steps are accessible (depending on the object) via specific functions (e.g., report_parameters).

Output

The report() function generates a report-object that contain in itself different representations (e.g., text, tables, plots). These different representations can be accessed via several functions, such as:

as.report_text(r): Detailed text.

as.report_text(r, summary=TRUE): Minimal text giving the minimal information.

as.report_table(r): Comprehensive table including most available indices.

as.report_table(r, summary=TRUE): Minimal table.

Note that for some report objects, some of these representations might be identical.

See also

Examples

library(report)

model <- t.test(mtcars$mpg ~ mtcars$am)
r <- report(model)

# Text
r
#> Effect sizes were labelled following Cohen's (1988) recommendations.
#> 
#> The Welch Two Sample t-test testing the difference of mtcars$mpg by mtcars$am
#> (mean in group 0 = 17.15, mean in group 1 = 24.39) suggests that the effect is
#> negative, statistically significant, and large (difference = -7.24, 95% CI
#> [-11.28, -3.21], t(18.33) = -3.77, p = 0.001; Cohen's d = -1.41, 95% CI [-2.26,
#> -0.53])
summary(r)
#> The Welch Two Sample t-test testing the difference of mtcars$mpg by mtcars$am
#> (mean in group 0 = 17.15, mean in group 1 = 24.39) suggests that the effect is
#> negative, statistically significant, and large (difference = -7.24, 95% CI
#> [-11.28, -3.21], t(18.33) = -3.77, p = 0.001, Cohen's d = -1.41)

# Tables
as.data.frame(r)
#> Welch Two Sample t-test
#> 
#> Parameter  |     Group | Mean_Group1 | Mean_Group2 | Difference
#> ---------------------------------------------------------------
#> mtcars$mpg | mtcars$am |       17.15 |       24.39 |      -7.24
#> 
#> Parameter  |          95% CI | t(18.33) |     p | Cohen's d |  Cohen's d  CI
#> ----------------------------------------------------------------------------
#> mtcars$mpg | [-11.28, -3.21] |    -3.77 | 0.001 |     -1.41 | [-2.26, -0.53]
#> 
#> Alternative hypothesis: two.sided
summary(as.data.frame(r))
#> Difference |          95% CI | t(18.33) |     p | Cohen's d |  Cohen's d  CI
#> ----------------------------------------------------------------------------
#> -7.24      | [-11.28, -3.21] |    -3.77 | 0.001 |     -1.41 | [-2.26, -0.53]
#> 
#> Alternative hypothesis: two.sided