report: Automated Results Reporting as a Practical Tool to Improve Reproducibility and Methodological Best Practices Adoption

report’s primary goal is to bridge the gap between R’s output and the formatted results contained in your manuscript. It automatically produces reports of models and data frames according to best practices guidelines (e.g., APA’s style), ensuring standardization and quality in results reporting.

Details

report-package

See also

Author

Maintainer: Rémi Thériault remi.theriault@mail.mcgill.ca (ORCID)

Authors:

Dominique Makowski dom.makowski@gmail.com (ORCID)

Daniel Lüdecke d.luedecke@uke.de (ORCID)

Indrajeet Patil patilindrajeet.science@gmail.com (ORCID)

Mattan S. Ben-Shachar matanshm@post.bgu.ac.il (ORCID)

Brenton M. Wiernik brenton@wiernik.org (ORCID)

Other contributors:

Rudolf Siegel mutlusun@users.noreply.github.com (ORCID) [contributor]

Camden Bock camden.bock@maine.edu (ORCID) [contributor]