Models' Bayes factor Report

Create a report of Bayes factors for model comparison.

# S3 method for bayesfactor_models
report(model, interpretation = "jeffreys1961", ...)

# S3 method for bayesfactor_inclusion
report(model, interpretation = "jeffreys1961", ...)

Arguments

model
Object of class bayesfactor_inclusion.

interpretation
Effect size interpretation set of rules (see interpret_bf).

...
Arguments passed to or from other methods.

See also

Examples

library(bayestestR)
library(report)

mo0 <- lm(Sepal.Length ~ 1, data = iris)
mo1 <- lm(Sepal.Length ~ Species, data = iris)
mo2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
mo3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)

# Bayes factor - models
BFmodels <- bayesfactor_models(mo1, mo2, mo3, denominator = mo0)

r <- report(BFmodels)
r

#> Bayes factors were computed using the BIC approximation, by which BF10 = exp((BIC0 - BIC1)/2). Compared to the (Intercept only) model (the least supported model), we found extreme evidence (BF = 1.69585e+29) in favour of the Species model; extreme evidence (BF = 5.84311e+55) in favour of the Species + Petal.Length model (the most supported model); extreme evidence (BF = 2.20324e+54) in favour of the Species * Petal.Length model.

table_short(r)

#> Model                                | Bayes factor
#> ---------------------------------------------------
#> [1] Species                          |     1.70e+29
#> [2] Species + Petal.Length           |     5.84e+55
#> [3] Species * Petal.Length           |     2.20e+54
#> [4] (Intercept only)                 |            1
#>                                      |             
#> Bayes Factor Type: BIC approximation |             
#> Against denominator - model 4        |             

# Bayes factor - inclusion
inc_bf <- bayesfactor_inclusion(BFmodels, prior_odds = c(1, 2, 3), match_models = TRUE)

r <- report(inc_bf)
r

#> Bayesian model averaging (BMA) was used to obtain the average evidence for each predictor. Since each model has a prior probability, it is possible to sum the prior probability of all models that include a predictor of interest (the prior inclusion probability), and of all models that do not include that predictor (the prior exclusion probability). After the data are observed, we can similarly consider the sums of the posterior models' probabilities to obtain the posterior inclusion probability and the posterior exclusion probability. The change from prior to posterior inclusion odds is the Inclusion Bayes factor. We found extreme evidence (BF = 3.89540e+55) in favour of including Species, with models including Species having an overall posterior probability of 95%; extreme evidence (BF = 6.89106e+26) in favour of including Petal.Length, with models including Petal.Length having an overall posterior probability of 95%; strong evidence (BF = 0.04) against including Petal.Length:Species, with models including Petal.Length:Species having an overall posterior probability of 5%.

table_short(r)

#> Terms                                | Pr(prior) | Pr(posterior) | Inclusion BF
#> -------------------------------------------------------------------------------
#> Species                              |      0.43 |          0.95 |     3.90e+55
#> Petal.Length                         |      0.29 |          0.95 |     6.89e+26
#> Petal.Length:Species                 |      0.43 |          0.05 |         0.04
#>                                      |           |               |             
#> Across matched models only,          |           |               |             
#> with custom prior odds of [1, 2, 3]. |           |               |