Data frame Report

Create a report of a data frame.

# S3 method for character
report(
  model,
  n_entries = 3,
  levels_percentage = FALSE,
  missing_percentage = FALSE,
  ...
)

# S3 method for data.frame
report(
  model,
  centrality = "mean",
  dispersion = TRUE,
  range = TRUE,
  distribution = FALSE,
  levels_percentage = FALSE,
  n_entries = 3,
  missing_percentage = FALSE,
  ...
)

# S3 method for factor
report(model, levels_percentage = FALSE, ...)

# S3 method for numeric
report(
  model,
  centrality = "mean",
  dispersion = TRUE,
  range = TRUE,
  distribution = FALSE,
  missing_percentage = FALSE,
  ...
)

Arguments

model
A data.frame or a vector.

n_entries
Number of different character entries to show. Can be "all".

levels_percentage
Show characters entries and factor levels by number (default) or percentage.

missing_percentage
Show missings by number (default) or percentage.

...
Arguments passed to or from other methods.

centrality
Character vector, indicating the index of centrality (either "mean" or "median").

dispersion
Show index of dispersion (sd if centrality = "mean", or mad if centrality = "median").

range
Show range.

distribution
Returns kurtosis and skewness in table.

Value

A list-object of class report, which contains further list-objects with a short and long description of the model summary, as well as a short and long table of parameters and fit indices.

See also

Examples

x <- c("A", "B", "C", "A", "B", "B", "D", "E", "B", "D", "A")
report(x)

#> 5 entries: B, n = 4; A, n = 3; D, n = 2 and 2 others (0 missing)

model_table(x)

#> n_Entries | n_Obs | n_Missing | percentage_Missing |                      Entries
#> ---------------------------------------------------------------------------------
#> 5.00      |    11 |         0 |               0.00 | B, n = 4; A, n = 3; D, n = 2

model_text(x)

#> 5 entries: B, n = 4; A, n = 3; D, n = 2 and 2 others (0 missing)

summary(model_table(x))

#> n_Entries | n_Obs | n_Missing |                      Entries
#> ------------------------------------------------------------
#> 5.00      |    11 |         0 | B, n = 4; A, n = 3; D, n = 2

summary(model_text(x))

#> 5 entries: B, n = 4; A, n = 3; D, n = 2 and 2 others

library(report)

r <- report(iris, centrality = "median", dispersion = FALSE,
            distribution = TRUE, missing_percentage = TRUE)
r

#> The data contains 150 observations of the following variables:
#> 
#>   - Sepal.Length: Mean = 5.84, SD = 0.83, Median = 5.80, MAD = 1.04, range: [4.30, 7.90], Skewness = 0.31, Kurtosis = -0.57, 0% missing
#>   - Sepal.Width: Mean = 3.06, SD = 0.44, Median = 3.00, MAD = 0.44, range: [2, 4.40], Skewness = 0.32, Kurtosis = 0.18, 0% missing
#>   - Petal.Length: Mean = 3.76, SD = 1.77, Median = 4.35, MAD = 1.85, range: [1, 6.90], Skewness = -0.27, Kurtosis = -1.40, 0% missing
#>   - Petal.Width: Mean = 1.20, SD = 0.76, Median = 1.30, MAD = 1.04, range: [0.10, 2.50], Skewness = -0.10, Kurtosis = -1.34, 0% missing
#>   - Species: 3 levels: setosa (n = 50, 33.33%); versicolor (n = 50, 33.33%) and virginica (n = 50, 33.33%)

summary(r)

#> The data contains 150 observations of the following variables:
#> 
#>   - Sepal.Length: Median = 5.80, range = [4.30, 7.90], Skewness = 0.31, Kurtosis = -0.57, 0% missing
#>   - Sepal.Width: Median = 3.00, range = [2, 4.40], Skewness = 0.32, Kurtosis = 0.18, 0% missing
#>   - Petal.Length: Median = 4.35, range = [1, 6.90], Skewness = -0.27, Kurtosis = -1.40, 0% missing
#>   - Petal.Width: Median = 1.30, range = [0.10, 2.50], Skewness = -0.10, Kurtosis = -1.34, 0% missing
#>   - Species: 3 levels: setosa, n = 50; versicolor, n = 50 and virginica, n = 50

as.data.frame(r)

#> Variable     |      Level | n_Obs | Mean |   SD | Median |  MAD |  Min |  Max | Skewness | Kurtosis | percentage_Missing | percentage_Obs
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> Sepal.Length |            |   150 | 5.84 | 0.83 |   5.80 | 1.04 | 4.30 | 7.90 |     0.31 |    -0.57 |               0.00 |               
#> Sepal.Width  |            |   150 | 3.06 | 0.44 |   3.00 | 0.44 | 2.00 | 4.40 |     0.32 |     0.18 |               0.00 |               
#> Petal.Length |            |   150 | 3.76 | 1.77 |   4.35 | 1.85 | 1.00 | 6.90 |    -0.27 |    -1.40 |               0.00 |               
#> Petal.Width  |            |   150 | 1.20 | 0.76 |   1.30 | 1.04 | 0.10 | 2.50 |    -0.10 |    -1.34 |               0.00 |               
#> Species      |     setosa |    50 |      |      |        |      |      |      |          |          |                    |          33.33
#> Species      | versicolor |    50 |      |      |        |      |      |      |          |          |                    |          33.33
#> Species      |  virginica |    50 |      |      |        |      |      |      |          |          |                    |          33.33

summary(as.data.frame(r))

#>    Variable                Level       n_Obs            Mean      
#>  Length:7           setosa    :1   Min.   : 50.0   Min.   :1.199  
#>  Class :character   versicolor:1   1st Qu.: 50.0   1st Qu.:2.593  
#>  Mode  :character   virginica :1   Median :150.0   Median :3.408  
#>                     NA's      :4   Mean   :107.1   Mean   :3.465  
#>                                    3rd Qu.:150.0   3rd Qu.:4.279  
#>                                    Max.   :150.0   Max.   :5.843  
#>                                                    NA's   :3      
#>        SD             Median           MAD              Min       
#>  Min.   :0.4359   Min.   :1.300   Min.   :0.4448   Min.   :0.100  
#>  1st Qu.:0.6806   1st Qu.:2.575   1st Qu.:0.8896   1st Qu.:0.775  
#>  Median :0.7952   Median :3.675   Median :1.0378   Median :1.500  
#>  Mean   :0.9479   Mean   :3.612   Mean   :1.0934   Mean   :1.850  
#>  3rd Qu.:1.0624   3rd Qu.:4.713   3rd Qu.:1.2417   3rd Qu.:2.575  
#>  Max.   :1.7653   Max.   :5.800   Max.   :1.8533   Max.   :4.300  
#>  NA's   :3        NA's   :3       NA's   :3        NA's   :3      
#>       Max           Skewness           Kurtosis       percentage_Missing
#>  Min.   :2.500   Min.   :-0.27488   Min.   :-1.3955   Min.   :0         
#>  1st Qu.:3.925   1st Qu.:-0.14595   1st Qu.:-1.3509   1st Qu.:0         
#>  Median :5.650   Median : 0.10597   Median :-0.9548   Median :0         
#>  Mean   :5.425   Mean   : 0.06401   Mean   :-0.7810   Mean   :0         
#>  3rd Qu.:7.150   3rd Qu.: 0.31592   3rd Qu.:-0.3849   3rd Qu.:0         
#>  Max.   :7.900   Max.   : 0.31897   Max.   : 0.1810   Max.   :0         
#>  NA's   :3       NA's   :3          NA's   :3         NA's   :3         
#>  percentage_Obs 
#>  Min.   :33.33  
#>  1st Qu.:33.33  
#>  Median :33.33  
#>  Mean   :33.33  
#>  3rd Qu.:33.33  
#>  Max.   :33.33  
#>  NA's   :4      

if (require("dplyr")) {
  r <- iris %>%
    dplyr::group_by(Species) %>%
    report()
  r
  summary(r)
  as.data.frame(r)
  summary(as.data.frame(r))
}

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#>     Group             Variable             n_Obs         Mean      
#>  Length:12          Length:12          Min.   :50   Min.   :0.246  
#>  Class :character   Class :character   1st Qu.:50   1st Qu.:1.885  
#>  Mode  :character   Mode  :character   Median :50   Median :3.201  
#>                                        Mean   :50   Mean   :3.465  
#>                                        3rd Qu.:50   3rd Qu.:5.143  
#>                                        Max.   :50   Max.   :6.588  
#>        SD             Median           MAD              Min       
#>  Min.   :0.1054   Min.   :0.200   Min.   :0.0000   Min.   :0.100  
#>  1st Qu.:0.2554   1st Qu.:1.875   1st Qu.:0.2780   1st Qu.:1.300  
#>  Median :0.3375   Median :3.200   Median :0.2965   Median :2.250  
#>  Mean   :0.3578   Mean   :3.458   Mean   :0.3521   Mean   :2.633  
#>  3rd Qu.:0.4815   3rd Qu.:5.138   3rd Qu.:0.5189   3rd Qu.:4.350  
#>  Max.   :0.6359   Max.   :6.500   Max.   :0.6672   Max.   :4.900  
#>       Max           Skewness           Kurtosis          n_Missing
#>  Min.   :0.600   Min.   :-0.60651   Min.   :-0.66135   Min.   :0  
#>  1st Qu.:2.350   1st Qu.:-0.05575   1st Qu.:-0.45816   1st Qu.:0  
#>  Median :4.100   Median : 0.10589   Median :-0.17221   Median :0  
#>  Mean   :4.258   Mean   : 0.12752   Mean   : 0.04517   Mean   :0  
#>  3rd Qu.:6.075   3rd Qu.: 0.18155   3rd Qu.: 0.57588   3rd Qu.:0  
#>  Max.   :7.900   Max.   : 1.25386   Max.   : 1.43432   Max.   :0  

x <- factor(rep(c("A", "B", "C"), 10))
report(x)

#> 3 levels: A (n = 10, 33.33%); B (n = 10, 33.33%) and C (n = 10, 33.33%)

model_table(x)

#> Level | n_Obs | percentage_Obs
#> ------------------------------
#> A     |    10 |          33.33
#> B     |    10 |          33.33
#> C     |    10 |          33.33

model_text(x)

#> 3 levels: A (n = 10, 33.33%); B (n = 10, 33.33%) and C (n = 10, 33.33%)

summary(model_table(x))

#> Level | n_Obs
#> -------------
#> A     |    10
#> B     |    10
#> C     |    10

summary(model_text(x))

#> 3 levels: A, n = 10; B, n = 10 and C, n = 10

x <- rnorm(1000)
report(x)

#> Mean = 0.00, SD = 1.00, Median = 0.01, MAD = 1.00, range: [-2.94, 2.76], Skewness = 0.00, Kurtosis = -0.20, 0 missing

report(x, centrality = "median", missing_percentage = TRUE, distribution = TRUE)

#> Mean = 0.00, SD = 1.00, Median = 0.01, MAD = 1.00, range: [-2.94, 2.76], Skewness = 0.00, Kurtosis = -0.20, 0% missing

model_table(x)

#> Mean |   SD | Median |  MAD |   Min |  Max | n_Obs | Skewness | Kurtosis | n_Missing
#> ------------------------------------------------------------------------------------
#> 0.00 | 1.00 |   0.01 | 1.00 | -2.94 | 2.76 |  1000 |     0.00 |    -0.20 |         0

model_text(x)

#> Mean = 0.00, SD = 1.00, Median = 0.01, MAD = 1.00, range: [-2.94, 2.76], Skewness = 0.00, Kurtosis = -0.20, 0 missing

summary(model_table(x))

#> Mean |   SD |   Min |  Max | n_Obs | n_Missing
#> ----------------------------------------------
#> 0.00 | 1.00 | -2.94 | 2.76 |  1000 |         0

summary(model_text(x))

#> Mean = 0.00, SD = 1.00, range = [-2.94, 2.76], 0 missing