(General) Linear Models Report

Create a report of a (general) linear model.

# S3 method for lm
report(
  model,
  interpretation = "default",
  ci = 0.95,
  standardize = "refit",
  standardize_robust = FALSE,
  bootstrap = FALSE,
  iterations = 500,
  performance_metrics = "all",
  ...
)

Arguments

model
Model object.

interpretation
Interpret the standardized parameters using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

standardize_robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors.

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also parameters_bootstrap()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

performance_metrics
See model_performance.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

Examples

library(report)

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
text_short(r)

#> We fitted a linear model to predict Sepal.Length with Petal.Length and Species.The model's explanatory power is substantial (R2 = 0.84, adj. R2 = 0.83). The model's intercept is at 4.21. Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and not significant (beta = 1.16, 95% CI [-0.01, 2.32], std. beta = 1.16, p = 0.052).
#>   - The effect of Speciesversicolor is negative and can be considered as medium and not significant (beta = -0.88, 95% CI [-2.41, 0.65], std. beta = 0.61, p = 0.256).
#>   - The effect of Speciesvirginica is negative and can be considered as very large and significant (beta = -1.75, 95% CI [-3.32, -0.18], std. beta = 0.97, p < .05).
#>   - The effect of Petal.Length:Speciesversicolor is positive and can be considered as very large and not significant (beta = 0.61, 95% CI [-0.63, 1.85], std. beta = -0.88, p = 0.334).
#>   - The effect of Petal.Length:Speciesvirginica is positive and can be considered as very large and not significant (beta = 0.97, 95% CI [-0.26, 2.19], std. beta = -1.75, p = 0.120).

text_long(r)

#> We fitted a linear model (estimated using OLS) to predict Sepal.Length with Petal.Length and Species (formula = Sepal.Length ~ Petal.Length * Species). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Funder's (2019) recommendations.
#> 
#> The model explains a significant and substantial proportion of variance (R2 = 0.84, F(5, 144) = 151.71, p < .001, adj. R2 = 0.83). The model's intercept, corresponding to Sepal.Length = 0, Petal.Length = 0 and Species = setosa, is at 4.21 (SE = 0.41, 95% CI [3.41, 5.02], p < .001). Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and not significant (beta = 1.16, SE = 0.59, 95% CI [-0.01, 2.32], std. beta = 1.16, p = 0.052).
#>   - The effect of Speciesversicolor is negative and can be considered as medium and not significant (beta = -0.88, SE = 0.77, 95% CI [-2.41, 0.65], std. beta = 0.61, p = 0.256).
#>   - The effect of Speciesvirginica is negative and can be considered as very large and significant (beta = -1.75, SE = 0.79, 95% CI [-3.32, -0.18], std. beta = 0.97, p < .05).
#>   - The effect of Petal.Length:Speciesversicolor is positive and can be considered as very large and not significant (beta = 0.61, SE = 0.63, 95% CI [-0.63, 1.85], std. beta = -0.88, p = 0.334).
#>   - The effect of Petal.Length:Speciesvirginica is positive and can be considered as very large and not significant (beta = 0.97, SE = 0.62, 95% CI [-0.26, 2.19], std. beta = -1.75, p = 0.120).

table_short(r)

#> Parameter                      | Coefficient | CI_low | CI_high |    p | Std_Coefficient |  Fit
#> -----------------------------------------------------------------------------------------------
#> (Intercept)                    |        0.49 |  -1.03 |    2.01 | 0.52 |            0.49 |     
#> Petal.Length                   |        1.16 |  -0.01 |    2.32 | 0.05 |            1.16 |     
#> Speciesversicolor              |       -0.88 |  -2.41 |    0.65 | 0.26 |            0.61 |     
#> Speciesvirginica               |       -1.75 |  -3.32 |   -0.18 | 0.03 |            0.97 |     
#> Petal.Length:Speciesversicolor |        0.61 |  -0.63 |    1.85 | 0.33 |           -0.88 |     
#> Petal.Length:Speciesvirginica  |        0.97 |  -0.26 |    2.19 | 0.12 |           -1.75 |     
#>                                |             |        |         |      |                 |     
#> R2                             |             |        |         |      |                 | 0.84
#> R2 (adj.)                      |             |        |         |      |                 | 0.83

table_long(r)

#> Parameter                      | Coefficient |   SE | CI_low | CI_high |     t | df_error |    p | Std_Coefficient |    Fit
#> ---------------------------------------------------------------------------------------------------------------------------
#> (Intercept)                    |        0.49 | 0.77 |  -1.03 |    2.01 |  0.64 |      144 | 0.52 |            0.49 |       
#> Petal.Length                   |        1.16 | 0.59 |  -0.01 |    2.32 |  1.96 |      144 | 0.05 |            1.16 |       
#> Speciesversicolor              |       -0.88 | 0.77 |  -2.41 |    0.65 | -1.14 |      144 | 0.26 |            0.61 |       
#> Speciesvirginica               |       -1.75 | 0.79 |  -3.32 |   -0.18 | -2.20 |      144 | 0.03 |            0.97 |       
#> Petal.Length:Speciesversicolor |        0.61 | 0.63 |  -0.63 |    1.85 |  0.97 |      144 | 0.33 |           -0.88 |       
#> Petal.Length:Speciesvirginica  |        0.97 | 0.62 |  -0.26 |    2.19 |  1.56 |      144 | 0.12 |           -1.75 |       
#>                                |             |      |        |         |       |          |      |                 |       
#> AIC                            |             |      |        |         |       |          |      |                 | 106.77
#> BIC                            |             |      |        |         |       |          |      |                 | 127.84
#> R2                             |             |      |        |         |       |          |      |                 |   0.84
#> R2 (adj.)                      |             |      |        |         |       |          |      |                 |   0.83
#> RMSE                           |             |      |        |         |       |          |      |                 |   0.33

model <- glm(vs ~ disp, data = mtcars, family = "binomial")
r <- report(model)

#> Waiting for profiling to be done...

#> Waiting for profiling to be done...

#> Waiting for profiling to be done...

text_short(r)

#> We fitted a logistic model to predict vs with disp.The model's explanatory power is substantial (Tjur's R2 = 0.53). The model's intercept is at 4.14. Within this model:
#> 
#>   - The effect of disp is negative and can be considered as large and significant (beta = -2.68, 95% CI [-4.90, -1.27], std. beta = -2.68, p < .01).

text_long(r)

#> We fitted a logistic model (estimated using ML) to predict vs with disp (formula = vs ~ disp). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Chen's (2010) recommendations.The model's explanatory power is substantial (Tjur's R2 = 0.53). The model's intercept, corresponding to vs = 0 and disp = 0, is at 4.14 (SE = 1.39, 95% CI [1.81, 7.44], p < .01). Within this model:
#> 
#>   - The effect of disp is negative and can be considered as large and significant (beta = -2.68, SE = 0.88, 95% CI [-4.90, -1.27], std. beta = -2.68, p < .01).

table_short(r)

#> Parameter   | Coefficient | CI_low | CI_high |    p | Std_Coefficient |  Fit
#> ----------------------------------------------------------------------------
#> (Intercept) |       -0.85 |  -2.42 |    0.27 | 0.19 |           -0.85 |     
#> disp        |       -2.68 |  -4.90 |   -1.27 | 0.00 |           -2.68 |     
#>             |             |        |         |      |                 |     
#> Tjur's R2   |             |        |         |      |                 | 0.53

table_long(r)

#> Parameter       | Coefficient |   SE | CI_low | CI_high |     z | df_error |    p | Std_Coefficient |    Fit
#> ------------------------------------------------------------------------------------------------------------
#> (Intercept)     |       -0.85 | 0.64 |  -2.42 |    0.27 | -1.31 |       30 | 0.19 |           -0.85 |       
#> disp            |       -2.68 | 0.88 |  -4.90 |   -1.27 | -3.03 |       30 | 0.00 |           -2.68 |       
#>                 |             |      |        |         |       |          |      |                 |       
#> AIC             |             |      |        |         |       |          |      |                 |  26.70
#> BIC             |             |      |        |         |       |          |      |                 |  29.63
#> Tjur's R2       |             |      |        |         |       |          |      |                 |   0.53
#> RMSE            |             |      |        |         |       |          |      |                 |   0.84
#> LOGLOSS         |             |      |        |         |       |          |      |                 |   0.35
#> SCORE_LOG       |             |      |        |         |       |          |      |                 | -11.07
#> SCORE_SPHERICAL |             |      |        |         |       |          |      |                 |   0.10
#> PCP             |             |      |        |         |       |          |      |                 |   0.77