Model description

Model textual description.

report_intercept(model, parameters, ...)

report_model(model, ...)

# S3 method for glm
report_model(
  model,
  parameters = NULL,
  ci = NULL,
  ci_method = NULL,
  standardize = "refit",
  standardize_robust = FALSE,
  interpretation = NULL,
  bootstrap = FALSE,
  iterations = 500,
  test = NULL,
  rope_range = NULL,
  rope_ci = NULL,
  df_method = NULL,
  ...
)

report_parameters(model, parameters, prefix = "  - ", ...)

report_performance(model, performance = NULL, ...)

Arguments

model
Object.

parameters
A parameters table obtained via parameters::model_parameters().

...
Arguments passed to or from other methods.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi), "ETI" (see eti) or "SI" (see si).

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

standardize_robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors.

interpretation
Interpret the standardized parameters using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests. For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

df_method
Method for computing degrees of freedom for p values, standard errors and confidence intervals (CI). May be "wald" (default, see degrees_of_freedom), "ml1" (see dof_ml1), "betwithin" (see dof_betwithin), "satterthwaite" (see dof_satterthwaite) or "kenward" (see dof_kenward). Note that when df_method is not "wald", robust standard errors etc. cannot be computed.

prefix
Prefix character that appears in front of each line.

performance
A performance table obtained via performance::model_performance().

See also

Examples

model <- lm(Sepal.Length ~ Species, data = iris)
report_intercept(model)

#>  The model's intercept, corresponding to Sepal.Length = 0 and Species = setosa, is at 5.01 (SE = 0.07, 95% CI [4.86, 5.15], p < .001).

library(lme4)
model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
report_intercept(model)

#>  The model's intercept, corresponding to Sepal.Length = 0, Petal.Length = 0 and Species = setosa, is at 2.50 (SE = 0.67, 95% CI [1.20, 3.81], p < .001).

model <- lm(Sepal.Length ~ Species, data = iris)
report_model(model)

#> We fitted a linear model (estimated using OLS) to predict Sepal.Length with Species (formula = Sepal.Length ~ Species). Standardized parameters were obtained by fitting the model on a standardized version of the dataset.

library(lme4)
model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
report_model(model)

#> We fitted a linear mixed model (estimated using REML and nloptwrap optimizer) to predict Sepal.Length with Petal.Length (formula = Sepal.Length ~ Petal.Length). The model included Species as random effects (formula = ~1 | Species). Standardized parameters were obtained by fitting the model on a standardized version of the dataset.

model <- lm(Sepal.Length ~ Species, data = iris)
report_parameters(model)

#>   - The effect of Species [versicolor] is positive and can be considered as significant (beta = 0.93, SE = 0.10, 95% CI [0.73, 1.13], p < .001).
#>   - The effect of Species [virginica] is positive and can be considered as significant (beta = 1.58, SE = 0.10, 95% CI [1.38, 1.79], p < .001).

library(lme4)
model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
report_parameters(model)

#>   - The effect of Petal.Length is positive and can be considered as significant (beta = 0.89, SE = 0.06, 95% CI [0.76, 1.01], p < .001).

model <- lm(Sepal.Length ~ Species, data = iris)
report_performance(model)

#> 
#> 
#> The model explains a significant and substantial proportion of variance (R2 = 0.62, F(2, 147) = 119.26, p < .001, adj. R2 = 0.61).

library(lme4)
model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
report_performance(model)

#> The model's total explanatory power is substantial (conditional R2 = 0.97) and the part related to the fixed effects alone (marginal R2) is of 0.66.