Reporting the participant data

A helper function to help you format the participants data (age, sex, ...) in the participants section.

report_participants(
  data,
  age = NULL,
  sex = NULL,
  education = NULL,
  participants = NULL,
  group = NULL,
  spell_n = FALSE,
  ...
)

Arguments

data
A dataframe.

age
The name of the column containing the age.

sex
The name of the column containing the sex. Note that classes should be some of c("Male", "M", "Female", "F").

education
The name of the column containing education information.

participants
The name of the participants' identifier column (for instance in the case of repeated measures).

group
A character vector indicating the name(s) of the column(s) used for stratified description.

spell_n
Fully spell the sample size ("Three participants" instead of "3 participants").

...
Arguments passed to or from other methods.

Examples

library(report)
data <- data.frame(
  "Age" = c(22, 23, 54, 21, 8, 42),
  "Sex" = c("F", "F", "M", "M", "M", "F")
)

report_participants(data, age = "Age", sex = "Sex")

#> [1] "6 participants (Mean age = 28.33, SD = 16.62, range = [8, 54]; 50.00% females)"

# Years of education (relative to high school graduation)
data$Education <- c(0, 8, -3, -5, 3, 5)
report_participants(data, age = "Age", sex = "Sex", education = "Education")

#> [1] "6 participants (Mean age = 28.33, SD = 16.62, range = [8, 54]; 50.00% females; Mean education = 1.33, SD = 4.93, range = [-5, 8])"

# Education as factor
data$Education2 <- c(
  "Bachelor", "PhD", "Highschool",
  "Highschool", "Bachelor", "Bachelor"
)
report_participants(data, age = "Age", sex = "Sex", education = "Education2")

#> [1] "6 participants (Mean age = 28.33, SD = 16.62, range = [8, 54]; 50.00% females; Education: Bachelor, 50.00%; Highschool, 33.33% and PhD, 16.67%)"

# Repeated measures data
data <- data.frame(
  "Age" = c(22, 22, 54, 54, 8, 8),
  "Sex" = c("F", "F", "M", "M", "F", "F"),
  "Participant" = c("S1", "S1", "s2", "s2", "s3", "s3")
)

report_participants(data, age = "Age", sex = "Sex", participants = "Participant")

#> [1] "3 participants (Mean age = 28.00, SD = 23.58, range = [8, 54]; 66.67% females)"

# Grouped data
data <- data.frame(
  "Age" = c(22, 22, 54, 54, 8, 8, 42, 42),
  "Sex" = c("F", "F", "M", "M", "F", "F", "M", "M"),
  "Participant" = c("S1", "S1", "s2", "s2", "s3", "s3", "s4", "s4"),
  "Condition" = c("A", "A", "A", "A", "B", "B", "B", "B")
)

report_participants(data,
  age = "Age",
  sex = "Sex",
  participants = "Participant",
  group = "Condition"
)

#> Warning: Variable `data[[age]]` contains only two different values. Consider converting it to a factor.

#> Warning: Need at least 3 complete observations for type-2-skewness. Using 'type="1"' now.

#> Warning: Variable `data[[age]]` contains only two different values. Consider converting it to a factor.

#> Warning: Need at least 3 complete observations for type-2-skewness. Using 'type="1"' now.

#> [1] "For the 'Condition - A' group: 2 participants (Mean age = 38.00, SD = 22.63, range = [22, 54]; 50.00% females) and for the 'Condition - B' group: 2 participants (Mean age = 25.00, SD = 24.04, range = [8, 42]; 50.00% females)"

# Spell sample size
paste(
  report_participants(data, participants = "Participant", spell_n = TRUE),
  "were recruited in the study by means of torture and coercion."
)

#> [1] "Four participants (Mean age = 31.50, SD = 20.49, range = [8, 54]; 50.00% females) were recruited in the study by means of torture and coercion."