Report a descriptive table

Creates tables to describe different objects (see list of supported objects in report).

report_table(x, ...)

Arguments

x
The R object that you want to report (see list of of supported objects above).

...
Arguments passed to or from other methods.

Value

A data.frame.

Examples

library(report)

# Miscellaneous
r <- report_table(sessionInfo())
r

#> Package    | Version |                                                                                                                                                                                                                                  Reference
#> -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> bayestestR | 0.8.0.1 | Makowski, D., Ben-Shachar, M., & Lüdecke, D. (2019). bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework. Journal of Open Source Software, 4(40), 1541. doi:10.21105/joss.01541
#> dplyr      |   1.0.2 |                                                           Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2020). dplyr: A Grammar of Data Manipulation. R package version 1.0.2. https://CRAN.R-project.org/package=dplyr
#> lavaan     |   0.6.7 |                                                                           Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL http://www.jstatsoft.org/v48/i02/.
#> lme4       |  1.1.26 |                                                  Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#> Matrix     |  1.2.18 |                                                                         Douglas Bates and Martin Maechler (2019). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.2-18. https://CRAN.R-project.org/package=Matrix
#> R          |   4.0.3 |                                                                      R Core Team (2020). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
#> Rcpp       |   1.0.5 |                                                                   Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18. URL http://www.jstatsoft.org/v40/i08/.
#> report     |   0.2.0 |     Makowski, D., Lüdecke, D., & Ben-Shachar, M.S. (2020). Automated reporting as a practical tool to improve reproducibility and methodological best practices adoption. CRAN. Available from https://github.com/easystats/report. doi: .
#> rstanarm   |  2.21.1 |                                                                          Goodrich B, Gabry J, Ali I & Brilleman S. (2020). rstanarm: Bayesian applied regression modeling via Stan. R package version 2.21.1 https://mc-stan.org/rstanarm.

summary(r)

#> Package    | Version
#> --------------------
#> bayestestR | 0.8.0.1
#> dplyr      |   1.0.2
#> lavaan     |   0.6.7
#> lme4       |  1.1.26
#> Matrix     |  1.2.18
#> R          |   4.0.3
#> Rcpp       |   1.0.5
#> report     |   0.2.0
#> rstanarm   |  2.21.1

# Data
report_table(iris$Sepal.Length)

#> Mean |   SD | Median |  MAD |  Min |  Max | n_Obs | Skewness | Kurtosis | percentage_Missing
#> --------------------------------------------------------------------------------------------
#> 5.84 | 0.83 |   5.80 | 1.04 | 4.30 | 7.90 |   150 |     0.31 |    -0.55 |               0.00

report_table(as.character(round(iris$Sepal.Length, 1)))

#> n_Entries | n_Obs | n_Missing | percentage_Missing
#> --------------------------------------------------
#> 35.00     |   150 |         0 |               0.00

report_table(iris$Species)

#> Level      | n_Obs | percentage_Obs
#> -----------------------------------
#> setosa     |    50 |          33.33
#> versicolor |    50 |          33.33
#> virginica  |    50 |          33.33

report_table(iris)

#> Variable     |      Level | n_Obs | percentage_Obs | Mean |   SD | Median |  MAD |  Min |  Max | Skewness | Kurtosis | percentage_Missing
#> -----------------------------------------------------------------------------------------------------------------------------------------
#> Sepal.Length |            |   150 |                | 5.84 | 0.83 |   5.80 | 1.04 | 4.30 | 7.90 |     0.31 |    -0.55 |               0.00
#> Sepal.Width  |            |   150 |                | 3.06 | 0.44 |   3.00 | 0.44 | 2.00 | 4.40 |     0.32 |     0.23 |               0.00
#> Petal.Length |            |   150 |                | 3.76 | 1.77 |   4.35 | 1.85 | 1.00 | 6.90 |    -0.27 |    -1.40 |               0.00
#> Petal.Width  |            |   150 |                | 1.20 | 0.76 |   1.30 | 1.04 | 0.10 | 2.50 |    -0.10 |    -1.34 |               0.00
#> Species      |     setosa |    50 |          33.33 |      |      |        |      |      |      |          |          |                   
#> Species      | versicolor |    50 |          33.33 |      |      |        |      |      |      |          |          |                   
#> Species      |  virginica |    50 |          33.33 |      |      |        |      |      |      |          |          |                   

# h-tests
report_table(t.test(mpg ~ am, data = mtcars))

#> Parameter | Group | Mean_Group1 | Mean_Group2 | Difference | t(18.33) |     p |              CI |                  Method |     d |           d CI
#> --------------------------------------------------------------------------------------------------------------------------------------------------
#> mpg       |    am |       17.15 |       24.39 |       7.24 |    -3.77 | 0.001 | [-11.28, -3.21] | Welch Two Sample t-test | -1.76 | [-2.82, -0.67]

report_table(cor.test(iris$Sepal.Length, iris$Sepal.Width))

#> Parameter1        |       Parameter2 |     r | t(148) |     p |        95% CI |  Method
#> ---------------------------------------------------------------------------------------
#> iris$Sepal.Length | iris$Sepal.Width | -0.12 |  -1.44 | 0.152 | [-0.27, 0.04] | Pearson

# ANOVAs
report_table(aov(Sepal.Length ~ Species, data=iris))

#> For one-way between subjects designs, partial eta squared is equvilant to eta squared.
#> Returning eta squared.

#> Parameter | Sum_Squares |  df | Mean_Square |      F |      p | Eta2 |  Eta2 90% CI
#> -----------------------------------------------------------------------------------
#> Species   |       63.21 |   2 |       31.61 | 119.26 | < .001 | 0.62 | [0.54, 0.68]
#> Residuals |       38.96 | 147 |        0.27 |        |        |      |             

# GLMs
report_table(lm(Sepal.Length ~ Petal.Length * Species, data = iris))

#> Warning: longer object length is not a multiple of shorter object length

#> Error in x$Parameter[x$Parameter == names(att$pretty_names[match_pretty_names])] <- att$pretty_names[match_pretty_names]: NAs are not allowed in subscripted assignments

report_table(glm(vs ~ disp, data = mtcars, family = "binomial"))

#> Warning: longer object length is not a multiple of shorter object length

#> Error in x$Parameter[x$Parameter == names(att$pretty_names[match_pretty_names])] <- att$pretty_names[match_pretty_names]: NAs are not allowed in subscripted assignments

# Mixed models
if(require("lme4")){
  model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
  report_table(model)
}

#> Warning: longer object length is not a multiple of shorter object length

#> Error in x$Parameter[x$Parameter == names(att$pretty_names[match_pretty_names])] <- att$pretty_names[match_pretty_names]: NAs are not allowed in subscripted assignments

# Bayesian models
if(require("rstanarm")){
  model <- stan_glm(Sepal.Length ~ Species, data = iris, refresh=0, iter=600)
  report_table(model, effectsize_method="basic")
}

#> Warning: longer object length is not a multiple of shorter object length

#> Error in x$Parameter[x$Parameter == names(att$pretty_names[match_pretty_names])] <- att$pretty_names[match_pretty_names]: NAs are not allowed in subscripted assignments

# Structural Equation Models (SEM)
if (require("lavaan")) {
  structure <- " ind60 =~ x1 + x2 + x3
                 dem60 =~ y1 + y2 + y3
                 dem60 ~ ind60 "
  model <- lavaan::sem(structure, data = PoliticalDemocracy)
  report_table(model)
}

#> Warning: longer object length is not a multiple of shorter object length

#> Error in x$Parameter[x$Parameter == names(att$pretty_names[match_pretty_names])] <- att$pretty_names[match_pretty_names]: NAs are not allowed in subscripted assignments