Report a textual description of an object

Creates text containing a description of the parameters of R objects (see list of supported objects in report).

report_text(x, table = NULL, ...)

Arguments

x
The R object that you want to report (see list of of supported objects above).

table
A table obtained via report_table(). If not provided, will run it.

...
Arguments passed to or from other methods.

Value

An object of class report_text.

Examples

library(report)

# Miscellaneous
r <- report_text(sessionInfo())
r
#> Analyses were conducted using the R Statistical language (version 4.1.0; R Core Team, 2021) on macOS Catalina 10.15.7, using the packages Rcpp (version 1.0.6; Dirk Eddelbuettel and Romain Francois, 2011), Matrix (version 1.3.3; Douglas Bates and Martin Maechler, 2021), lme4 (version 1.1.27; Douglas Bates et al., 2015), rstanarm (version 2.21.1; Goodrich B et al., 2020), dplyr (version 1.0.6; Hadley Wickham et al., 2021), performance (version 0.7.2; Lüdecke et al., 2021), bayestestR (version 0.10.0; Makowski et al., 2019), report (version 0.3.5; Makowski et al., 2020) and lavaan (version 0.6.8; Yves Rosseel, 2012).
#> 
#> References
#> ----------
#>   - Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18. URL https://www.jstatsoft.org/v40/i08/.
#>   - Douglas Bates and Martin Maechler (2021). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.3-3. https://CRAN.R-project.org/package=Matrix
#>   - Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#>   - Goodrich B, Gabry J, Ali I & Brilleman S. (2020). rstanarm: Bayesian applied regression modeling via Stan. R package version 2.21.1 https://mc-stan.org/rstanarm.
#>   - Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2021). dplyr: A Grammar of Data Manipulation. R package version 1.0.6. https://CRAN.R-project.org/package=dplyr
#>   - Lüdecke et al., (2021). performance: An R Package for Assessment, Comparison and Testing of Statistical Models. Journal of Open Source Software, 6(60), 3139. https://doi.org/10.21105/joss.03139
#>   - Makowski, D., Ben-Shachar, M., & Lüdecke, D. (2019). bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework. Journal of Open Source Software, 4(40), 1541. doi:10.21105/joss.01541
#>   - Makowski, D., Ben-Shachar, M.S., Patil, I. & Lüdecke, D. (2020). Automated Results Reporting as a Practical Tool to Improve Reproducibility and Methodological Best Practices Adoption. CRAN. Available from https://github.com/easystats/report. doi: .
#>   - R Core Team (2021). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
#>   - Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL https://www.jstatsoft.org/v48/i02/.
summary(r)
#> The analysis was done using the R Statistical language (v4.1.0; R Core Team, 2021) on macOS Catalina 10.15.7, using the packages Rcpp (v1.0.6), Matrix (v1.3.3), lme4 (v1.1.27), rstanarm (v2.21.1), dplyr (v1.0.6), performance (v0.7.2), bayestestR (v0.10.0), report (v0.3.5) and lavaan (v0.6.8).

# Data
report_text(iris$Sepal.Length)
#> iris$Sepal.Length: n = 150, Mean = 5.84, SD = 0.83, Median = 5.80, MAD = 1.04, range: [4.30, 7.90], Skewness = 0.31, Kurtosis = -0.55, 0% missing
report_text(as.character(round(iris$Sepal.Length, 1)))
#> as.character(round(iris$Sepal.Length, 1)): 35 entries, such as 5 (6.67%); 5.1 (6.00%); 6.3 (6.00%) and 32 others (0 missing)
report_text(iris$Species)
#> iris$Species: 3 levels, namely setosa (n = 50, 33.33%), versicolor (n = 50, 33.33%) and virginica (n = 50, 33.33%)
report_text(iris)
#> The data contains 150 observations of the following 5 variables:
#>   - Sepal.Length: n = 150, Mean = 5.84, SD = 0.83, Median = 5.80, MAD = 1.04, range: [4.30, 7.90], Skewness = 0.31, Kurtosis = -0.55, 0% missing
#>   - Sepal.Width: n = 150, Mean = 3.06, SD = 0.44, Median = 3.00, MAD = 0.44, range: [2, 4.40], Skewness = 0.32, Kurtosis = 0.23, 0% missing
#>   - Petal.Length: n = 150, Mean = 3.76, SD = 1.77, Median = 4.35, MAD = 1.85, range: [1, 6.90], Skewness = -0.27, Kurtosis = -1.40, 0% missing
#>   - Petal.Width: n = 150, Mean = 1.20, SD = 0.76, Median = 1.30, MAD = 1.04, range: [0.10, 2.50], Skewness = -0.10, Kurtosis = -1.34, 0% missing
#>   - Species: 3 levels, namely setosa (n = 50, 33.33%), versicolor (n = 50, 33.33%) and virginica (n = 50, 33.33%)

# h-tests
report_text(t.test(iris$Sepal.Width, iris$Sepal.Length))
#> Effect sizes were labelled following Cohen's (1988) recommendations.
#> 
#> The Welch Two Sample t-test testing the difference between iris$Sepal.Width and iris$Sepal.Length (mean of x = 3.06, mean of y = 5.84) suggests that the effect is negative, statistically significant, and large (difference = -2.79, 95% CI [-2.94, -2.64], t(225.68) = -36.46, p < .001; Cohen's d = -4.21, 95% CI [-4.66, -3.76])
report_text(cor.test(iris$Sepal.Width, iris$Sepal.Length))
#> Effect sizes were labelled following Funder's (2019) recommendations.
#> 
#> The Pearson's product-moment correlation between iris$Sepal.Width and iris$Sepal.Length is negative, statistically not significant, and small (r = -0.12, 95% CI [-0.27, 0.04], t(148) = -1.44, p = 0.152)

# ANOVA
r <- report_text(aov(Sepal.Length ~ Species, data = iris))
#> For one-way between subjects designs, partial eta squared is equivalent to eta squared.
#> Returning eta squared.
#> For one-way between subjects designs, partial eta squared is equivalent to eta squared.
#> Returning eta squared.
r
#> The ANOVA (formula: Sepal.Length ~ Species) suggests that:
#> 
#>   - The main effect of Species is statistically significant and large (F(2, 147) = 119.26, p < .001; Eta2 = 0.62, 90% CI [0.54, 0.68])
#> 
#> Effect sizes were labelled following Field's (2013) recommendations.
summary(r)
#> The ANOVA suggests that:
#> 
#>   - The main effect of Species is statistically significant and large (F(2, 147) = 119.26, p < .001, Eta2 = 0.62)

# GLMs
r <- report_text(lm(Sepal.Length ~ Petal.Length * Species, data = iris))
r
#> We fitted a linear model (estimated using OLS) to predict Sepal.Length with Petal.Length and Species (formula: Sepal.Length ~ Petal.Length * Species). The model explains a statistically significant and substantial proportion of variance (R2 = 0.84, F(5, 144) = 151.71, p < .001, adj. R2 = 0.83). The model's intercept, corresponding to Petal.Length = 0 and Species = setosa, is at 4.21 (95% CI [3.41, 5.02], t(144) = 10.34, p < .001). Within this model:
#> 
#>   - The effect of Petal.Length is statistically non-significant and positive (beta = 0.54, 95% CI [-4.76e-03, 1.09], t(144) = 1.96, p = 0.052; Std. beta = 1.16, 95% CI [-0.01, 2.32])
#>   - The effect of Species [versicolor] is statistically significant and negative (beta = -1.81, 95% CI [-2.99, -0.62], t(144) = -3.02, p < .01; Std. beta = -0.88, 95% CI [-2.41, 0.65])
#>   - The effect of Species [virginica] is statistically significant and negative (beta = -3.15, 95% CI [-4.41, -1.90], t(144) = -4.97, p < .001; Std. beta = -1.75, 95% CI [-3.32, -0.18])
#>   - The interaction effect of Species [versicolor] on Petal.Length is statistically non-significant and positive (beta = 0.29, 95% CI [-0.30, 0.87], t(144) = 0.97, p = 0.334; Std. beta = 0.61, 95% CI [-0.63, 1.85])
#>   - The interaction effect of Species [virginica] on Petal.Length is statistically non-significant and positive (beta = 0.45, 95% CI [-0.12, 1.03], t(144) = 1.56, p = 0.120; Std. beta = 0.97, 95% CI [-0.26, 2.19])
#> 
#> Standardized parameters were obtained by fitting the model on a standardized version of the dataset.
summary(r)
#> We fitted a linear model to predict Sepal.Length with Petal.Length and Species. The model's explanatory power is substantial (R2 = 0.84, adj. R2 = 0.83). The model's intercept is at 4.21 (95% CI [3.41, 5.02]). Within this model:
#> 
#>   - The effect of Petal.Length is statistically non-significant and positive (beta = 0.54, 95% CI [-4.76e-03, 1.09], t(144) = 1.96, p = 0.052, Std. beta = 1.16)
#>   - The effect of Species [versicolor] is statistically significant and negative (beta = -1.81, 95% CI [-2.99, -0.62], t(144) = -3.02, p < .01, Std. beta = -0.88)
#>   - The effect of Species [virginica] is statistically significant and negative (beta = -3.15, 95% CI [-4.41, -1.90], t(144) = -4.97, p < .001, Std. beta = -1.75)
#>   - The interaction effect of Species [versicolor] on Petal.Length is statistically non-significant and positive (beta = 0.29, 95% CI [-0.30, 0.87], t(144) = 0.97, p = 0.334, Std. beta = 0.61)
#>   - The interaction effect of Species [virginica] on Petal.Length is statistically non-significant and positive (beta = 0.45, 95% CI [-0.12, 1.03], t(144) = 1.56, p = 0.120, Std. beta = 0.97)
# \donttest{
if (require("lme4")) {
  model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
  r <- report_text(model)
  r
  summary(r)
}
#> We fitted a linear mixed model to predict Sepal.Length with Petal.Length. The model included Species as random effect. The model's total explanatory power is substantial (conditional R2 = 0.97) and the part related to the fixed effects alone (marginal R2) is of 0.66. The model's intercept is at 2.50 (95% CI [1.20, 3.81]). Within this model:
#> 
#>   - The effect of Petal.Length is statistically significant and positive (beta = 0.89, 95% CI [0.76, 1.01], t(146) = 13.93, p < .001, Std. beta = 1.89)

# Bayesian models
if (require("rstanarm")) {
  model <- stan_glm(mpg ~ cyl + wt, data = mtcars, refresh = 0, iter = 600)
  r <- report_text(model)
  r
  summary(r)
}
#> Possible multicollinearity between wt and cyl (r = 0.78). This might lead to inappropriate results. See 'Details' in '?rope'.
#> Possible multicollinearity between wt and cyl (r = 0.79). This might lead to inappropriate results. See 'Details' in '?rope'.
#> Possible multicollinearity between wt and cyl (r = 0.78). This might lead to inappropriate results. See 'Details' in '?rope'.
#> Possible multicollinearity between wt and cyl (r = 0.74). This might lead to inappropriate results. See 'Details' in '?rope'.
#> We fitted a Bayesian linear model to predict mpg with cyl and wt. Priors over parameters were set as normal (mean = 0.00, SD = 8.44) and normal (mean = 0.00, SD = 15.40) distributions. The model's explanatory power is substantial (R2 = 0.82, adj. R2 = 0.79). The model's intercept is at 39.74 (95% CI [36.50, 43.50]). Within this model:
#> 
#>   - The effect of cyl (Median = -1.50, 95% CI [-2.35, -0.62]) has 99.83%, 99.58% and 24.17% probability of being negative (< 0), significant (< -0.30) and large (< -1.81). The estimation successfully converged (Rhat = 1.001) but the indices are unreliable (ESS = 717)
#>   - The effect of wt (Median = -3.21, 95% CI [-4.91, -1.71]) has 100.00%, 100.00% and 95.50% probability of being negative (< 0), significant (< -0.30) and large (< -1.81). The estimation successfully converged (Rhat = 0.999) but the indices are unreliable (ESS = 704)
# }