Modify report output.

Functions to access different representations of a report-object. See the details section below.

text_long(r, ...)

text_short(r, ...)

table_long(r, ...)

table_short(r, ...)

Arguments

r
Object of class report.

...
Arguments passed to or from other methods.

Value

Character string or dataframe depending on the method.

Details

Dataframes and vectors

Correlations and t-tests (htest)

ANOVAs

Regression models (glm, lm, ...)

Mixed models (glmer, lmer, glmmTMB, ...)

Bayesian models (stanreg, brms...)

Change output type

The report() function generates a report-object that contain in itself different representations (e.g., text, tables, plots). These different representations can be accessed via several functions, such as:

text_long(): Detailed text.

text_short(): Minimal text giving the minimal information.

table_long(): Comprehensive table including most available indices.

table_short(): Minimal table.

Note that for some report objects, some of these representations might be identical.