Automatic reporting of R objects

Create reports of different objects. See the documentation for your object's class:

System and packages (sessionInfo)

Dataframes and vectors

Correlations and t-tests (htest)

ANOVAs (aov, anova, aovlist, ...)

Regression models (glm, lm, ...)

Mixed models (glmer, lmer, glmmTMB, ...)

Bayesian models (stanreg, brms...)

Bayes factors (from bayestestR)

Structural Equation Models (SEM) (from lavaan)

Model comparison (from performance())

Most of the time, the object created by the report() function can be further transformed, for instance summarized (using summary()), or converted to a table (using as.data.frame()).

Usage

report(x, ...)

Arguments

x

The R object that you want to report (see list of of supported objects above).

...

Arguments passed to or from other methods.

Value

A list-object of class report, which contains further list-objects with a short and long description of the model summary, as well as a short and long table of parameters and fit indices.

Details

Organization

report_table and report_text are the two distal representations of a report, and are the two provided in report(). However, intermediate steps are accessible (depending on the object) via specific functions (e.g., report_parameters).

Output

The report() function generates a report-object that contain in itself different representations (e.g., text, tables, plots). These different representations can be accessed via several functions, such as:

as.report_text(r): Detailed text.

as.report_text(r, summary=TRUE): Minimal text giving the minimal information.

as.report_table(r): Comprehensive table including most available indices.

as.report_table(r, summary=TRUE): Minimal table.

Note that for some report objects, some of these representations might be identical.

See also

Examples

library(report)

model <- t.test(mpg ~ am, data = mtcars)
r <- report(model)
#> Registered S3 methods overwritten by 'effectsize':
#>   method              from      
#>   standardize.Surv    datawizard
#>   standardize.bcplm   datawizard
#>   standardize.clm2    datawizard
#>   standardize.default datawizard
#>   standardize.mediate datawizard
#>   standardize.wbgee   datawizard
#>   standardize.wbm     datawizard
#> Warning: Unable to retrieve data from htest object. Using t_to_d() approximation.

# Text
r
#> 
#> 
#> The Welch Two Sample t-test testing the difference of mpg by am (mean in group 0 = 17.15, mean in group 1 = 24.39) suggests that the effect is negative, statistically significant, and large (difference = -7.24, 95% CI [-11.28, -3.21], t(18.33) = -3.77, p = 0.001; Cohen's d = -1.76, 95% CI [-2.82, -0.67])
summary(r)
#> The Welch Two Sample t-test testing the difference of mpg by am (mean in group 0 = 17.15, mean in group 1 = 24.39) suggests that the effect is negative, statistically significant, and large (difference = -7.24, 95% CI [-11.28, -3.21], t(18.33) = -3.77, p = 0.001, Cohen's d = -1.76)

# Tables
as.data.frame(r)
#> Welch Two Sample t-test
#> 
#> Parameter | Group | Mean_Group1 | Mean_Group2 | Difference |          95% CI | t(18.33) |     p |     d |          d  CI
#> ------------------------------------------------------------------------------------------------------------------------
#> mpg       |    am |       17.15 |       24.39 |      -7.24 | [-11.28, -3.21] |    -3.77 | 0.001 | -1.76 | [-2.82, -0.67]
#> 
#> Alternative hypothesis: two.sided
summary(as.data.frame(r))
#> Difference |          95% CI | t(18.33) |     p |     d |          d  CI
#> ------------------------------------------------------------------------
#> -7.24      | [-11.28, -3.21] |    -3.77 | 0.001 | -1.76 | [-2.82, -0.67]
#> 
#> Alternative hypothesis: two.sided