Reporting ANOVAs

Create reports for ANOVA models.

# S3 method for aov
report(x, ...)

# S3 method for aov
report_effectsize(x, ...)

# S3 method for aov
report_table(x, ...)

# S3 method for aov
report_statistics(x, table = NULL, ...)

# S3 method for aov
report_parameters(x, ...)

# S3 method for aov
report_model(x, table = NULL, ...)

# S3 method for aov
report_info(x, effectsize = NULL, ...)

# S3 method for aov
report_text(x, table = NULL, ...)

Arguments

x
Object of class aov, anova or aovlist.

...
Arguments passed to or from other methods.

table
Provide the output of report_table() to avoid its re-computation.

effectsize
Provide the output of report_effectsize() to avoid its re-computation.

Value

An object of class report.

See also

Examples

data <- iris
data$Cat1 <- rep(c("A", "B"), length.out = nrow(data))

model <- aov(Sepal.Length ~ Species * Cat1, data = data)
r <- report(model)
r

#> The ANOVA (formula: Sepal.Length ~ Species * Cat1) suggests that:
#> 
#>   - The main effect of Species is statistically significant and large (F(2, 144) = 118.43, p < .001; Eta2 (partial) = 0.62, 90% CI [0.54, 0.68])
#>   - The main effect of Cat1 is statistically not significant and very small (F(1, 144) = 6.25e-03, p = 0.937; Eta2 (partial) = 4.34e-05, 90% CI [0.00, 3.18e-03])
#>   - The interaction between Species and Cat1 is statistically not significant and small (F(2, 144) = 0.98, p = 0.377; Eta2 (partial) = 0.01, 90% CI [0.00, 0.05])
#> 
#> Effect sizes were labelled following Field's (2013) recommendations.

summary(r)

#> The ANOVA suggests that:
#> 
#>   - The main effect of Species is statistically significant and large (F(2, 144) = 118.43, p < .001, Eta2 (partial) = 0.62)
#>   - The main effect of Cat1 is statistically not significant and very small (F(1, 144) = 6.25e-03, p = 0.937, Eta2 (partial) = 4.34e-05)
#>   - The interaction between Species and Cat1 is statistically not significant and small (F(2, 144) = 0.98, p = 0.377, Eta2 (partial) = 0.01)

as.data.frame(r)

#> Parameter    | Sum_Squares |  df | Mean_Square |        F |      p | Eta2 (partial) | Eta2_partial 90% CI
#> ---------------------------------------------------------------------------------------------------------
#> Species      |       63.21 |   2 |       31.61 |   118.43 | < .001 |           0.62 |        [0.54, 0.68]
#> Cat1         |    1.67e-03 |   1 |    1.67e-03 | 6.25e-03 | 0.937  |       4.34e-05 |        [0.00, 0.00]
#> Species:Cat1 |        0.52 |   2 |        0.26 |     0.98 | 0.377  |           0.01 |        [0.00, 0.05]
#> Residuals    |       38.43 | 144 |        0.27 |          |        |                |                    

summary(as.data.frame(r))

#> Parameter    | Sum_Squares |  df | Mean_Square |        F |      p | Eta2 (partial)
#> -----------------------------------------------------------------------------------
#> Species      |       63.21 |   2 |       31.61 |   118.43 | < .001 |           0.62
#> Cat1         |    1.67e-03 |   1 |    1.67e-03 | 6.25e-03 | 0.937  |       4.34e-05
#> Species:Cat1 |        0.52 |   2 |        0.26 |     0.98 | 0.377  |           0.01
#> Residuals    |       38.43 | 144 |        0.27 |          |        |