Template to add report support for new objects

Template file to add report support for new objects. Check-out the vignette on Supporting New Models.

# S3 method for default
report(x, ...)

# S3 method for default
report_effectsize(x, ...)

# S3 method for default
report_table(x, ...)

# S3 method for default
report_statistics(x, ...)

# S3 method for default
report_parameters(x, ...)

# S3 method for default
report_intercept(x, ...)

# S3 method for default
report_model(x, ...)

# S3 method for default
report_random(x, ...)

# S3 method for default
report_priors(x, ...)

# S3 method for default
report_performance(x, ...)

# S3 method for default
report_info(x, ...)

# S3 method for default
report_text(x, ...)

Arguments

x
Object of class NEW OBJECT.

...
Arguments passed to or from other methods.

Value

A list-object of class report, which contains further list-objects with a short and long description of the model summary, as well as a short and long table of parameters and fit indices.

See also

Examples

library(report)

# Add a reproducible example instead of the following
model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
r

#> We fitted a linear model (estimated using OLS) to predict Sepal.Length with Petal.Length and Species (formula: Sepal.Length ~ Petal.Length * Species). The model explains a significant and substantial proportion of variance (R2 = 0.84, F(5, 144) = 151.71, p < .001, adj. R2 = 0.83). The model's intercept, corresponding to Petal.Length = 0 and Species = setosa, is at 4.21 (95% CI [3.41, 5.02], t(144) = 10.34, p < .001). Within this model:
#> 
#>   - The effect of Petal.Length is non-significantly positive (beta = 0.54, 95% CI [-4.76e-03, 1.09], t(144) = 1.96, p = 0.052; Std. beta = 1.16, 95% CI [-0.01, 2.32])
#>   - The effect of Species [versicolor] is significantly negative (beta = -1.81, 95% CI [-2.99, -0.62], t(144) = -3.02, p < .01; Std. beta = -0.88, 95% CI [-2.41, 0.65])
#>   - The effect of Species [virginica] is significantly negative (beta = -3.15, 95% CI [-4.41, -1.90], t(144) = -4.97, p < .001; Std. beta = -1.75, 95% CI [-3.32, -0.18])
#>   - The interaction effect of Species [versicolor] on Petal.Length is non-significantly positive (beta = 0.29, 95% CI [-0.30, 0.87], t(144) = 0.97, p = 0.334; Std. beta = 0.61, 95% CI [-0.63, 1.85])
#>   - The interaction effect of Species [virginica] on Petal.Length is non-significantly positive (beta = 0.45, 95% CI [-0.12, 1.03], t(144) = 1.56, p = 0.120; Std. beta = 0.97, 95% CI [-0.26, 2.19])
#> 
#> Standardized parameters were obtained by fitting the model on a standardized version of the dataset.

summary(r)

#> We fitted a linear model to predict Sepal.Length with Petal.Length and Species. The model's explanatory power is substantial (R2 = 0.84, adj. R2 = 0.83). The model's intercept is at 4.21 (95% CI [3.41, 5.02]). Within this model:
#> 
#>   - The effect of Petal.Length is non-significantly positive (beta = 0.54, 95% CI [-4.76e-03, 1.09], t(144) = 1.96, p = 0.052, Std. beta = 1.16)
#>   - The effect of Species [versicolor] is significantly negative (beta = -1.81, 95% CI [-2.99, -0.62], t(144) = -3.02, p < .01, Std. beta = -0.88)
#>   - The effect of Species [virginica] is significantly negative (beta = -3.15, 95% CI [-4.41, -1.90], t(144) = -4.97, p < .001, Std. beta = -1.75)
#>   - The interaction effect of Species [versicolor] on Petal.Length is non-significantly positive (beta = 0.29, 95% CI [-0.30, 0.87], t(144) = 0.97, p = 0.334, Std. beta = 0.61)
#>   - The interaction effect of Species [virginica] on Petal.Length is non-significantly positive (beta = 0.45, 95% CI [-0.12, 1.03], t(144) = 1.56, p = 0.120, Std. beta = 0.97)

as.data.frame(r)

#> Error in cat(insight::format_table(format(x, ...), ...)): argument 1 (type 'list') cannot be handled by 'cat'

summary(as.data.frame(r))

#> Error in cat(insight::format_table(format(x, ...), ...)): argument 1 (type 'list') cannot be handled by 'cat'