(General) Linear Models Report

Create a report of a (general) linear model.

# S3 method for default
report(
  model,
  interpretation = "default",
  ci = 0.95,
  standardize = "refit",
  standardize_robust = FALSE,
  bootstrap = FALSE,
  iterations = 500,
  performance_metrics = "all",
  ...
)

Arguments

model
Model object.

interpretation
Interpret the standardized parameters using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

standardize_robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors.

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

performance_metrics
See model_performance.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

Value

A list-object of class report, which contains further list-objects with a short and long description of the model summary, as well as a short and long table of parameters and fit indices.

See also

Examples

library(report)

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
text_short(r)

#> We fitted a linear model to predict Sepal.Length with Petal.Length and Species.The model's explanatory power is substantial (R2 = 0.84, adj. R2 = 0.83). The model's intercept is at 4.21. Within this model:
#> 
#>   - The effect of (Intercept) is positive and can be considered as medium and significant (beta = 4.21, 95% CI [3.41, 5.02], std. beta = 0.49, p < .001).
#>   - The effect of Petal.Length is positive and can be considered as very large and not significant (beta = 0.54, 95% CI [0.00, 1.09], std. beta = 1.16, p = 0.052).
#>   - The effect of Species [versicolor] is negative and can be considered as very large and significant (beta = -1.81, 95% CI [-2.99, -0.62], std. beta = -0.88, p < .01).
#>   - The effect of Species [virginica] is negative and can be considered as very large and significant (beta = -3.15, 95% CI [-4.41, -1.90], std. beta = -1.75, p < .001).
#>   - The interaction effect of Species [versicolor] on Petal.Length is positive and can be considered as medium and not significant (beta = 0.29, 95% CI [-0.30, 0.87], std. beta = 0.61, p = 0.334).
#>   - The interaction effect of Species [virginica] on Petal.Length is positive and can be considered as very large and not significant (beta = 0.45, 95% CI [-0.12, 1.03], std. beta = 0.97, p = 0.120).

text_long(r)

#> We fitted a linear model (estimated using OLS) to predict Sepal.Length with Petal.Length and Species (formula = Sepal.Length ~ Petal.Length * Species). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Funder's (2019) recommendations.
#> 
#> The model explains a significant and substantial proportion of variance (R2 = 0.84, F(5, 144) = 151.71, p < .001, adj. R2 = 0.83). The model's intercept, corresponding to Sepal.Length = 0, Petal.Length = 0 and Species = setosa, is at 4.21 (SE = 0.41, 95% CI [3.41, 5.02], p < .001). Within this model:
#> 
#>   - The effect of (Intercept) is positive and can be considered as medium and significant (beta = 4.21, SE = 0.41, 95% CI [3.41, 5.02], std. beta = 0.49, p < .001).
#>   - The effect of Petal.Length is positive and can be considered as very large and not significant (beta = 0.54, SE = 0.28, 95% CI [0.00, 1.09], std. beta = 1.16, p = 0.052).
#>   - The effect of Species [versicolor] is negative and can be considered as very large and significant (beta = -1.81, SE = 0.60, 95% CI [-2.99, -0.62], std. beta = -0.88, p < .01).
#>   - The effect of Species [virginica] is negative and can be considered as very large and significant (beta = -3.15, SE = 0.63, 95% CI [-4.41, -1.90], std. beta = -1.75, p < .001).
#>   - The interaction effect of Species [versicolor] on Petal.Length is positive and can be considered as medium and not significant (beta = 0.29, SE = 0.30, 95% CI [-0.30, 0.87], std. beta = 0.61, p = 0.334).
#>   - The interaction effect of Species [virginica] on Petal.Length is positive and can be considered as very large and not significant (beta = 0.45, SE = 0.29, 95% CI [-0.12, 1.03], std. beta = 0.97, p = 0.120).

table_short(r)

#> Parameter                           | Coefficient |             CI |      p | Coefficient (std.) |  Fit
#> -------------------------------------------------------------------------------------------------------
#> (Intercept)                         |        4.21 | [ 3.41,  5.02] | < .001 |               0.49 |     
#> Petal.Length                        |        0.54 | [ 0.00,  1.09] | 0.052  |               1.16 |     
#> Species [versicolor]                |       -1.81 | [-2.99, -0.62] | 0.003  |              -0.88 |     
#> Species [virginica]                 |       -3.15 | [-4.41, -1.90] | < .001 |              -1.75 |     
#> Petal.Length * Species [versicolor] |        0.29 | [-0.30,  0.87] | 0.334  |               0.61 |     
#> Petal.Length * Species [virginica]  |        0.45 | [-0.12,  1.03] | 0.120  |               0.97 |     
#>                                     |             |                |        |                    |     
#> R2                                  |             |                |        |                    | 0.84
#> R2 (adj.)                           |             |                |        |                    | 0.83

table_long(r)

#> Parameter                           | Coefficient |   SE |             CI |     t |  df |      p | Coefficient (std.) |    Fit
#> ------------------------------------------------------------------------------------------------------------------------------
#> (Intercept)                         |        4.21 | 0.41 | [ 3.41,  5.02] | 10.34 | 144 | < .001 |               0.49 |       
#> Petal.Length                        |        0.54 | 0.28 | [ 0.00,  1.09] |  1.96 | 144 | 0.052  |               1.16 |       
#> Species [versicolor]                |       -1.81 | 0.60 | [-2.99, -0.62] | -3.02 | 144 | 0.003  |              -0.88 |       
#> Species [virginica]                 |       -3.15 | 0.63 | [-4.41, -1.90] | -4.97 | 144 | < .001 |              -1.75 |       
#> Petal.Length * Species [versicolor] |        0.29 | 0.30 | [-0.30,  0.87] |  0.97 | 144 | 0.334  |               0.61 |       
#> Petal.Length * Species [virginica]  |        0.45 | 0.29 | [-0.12,  1.03] |  1.56 | 144 | 0.120  |               0.97 |       
#>                                     |             |      |                |       |     |        |                    |       
#> AIC                                 |             |      |                |       |     |        |                    | 106.77
#> BIC                                 |             |      |                |       |     |        |                    | 127.84
#> R2                                  |             |      |                |       |     |        |                    |   0.84
#> R2 (adj.)                           |             |      |                |       |     |        |                    |   0.83
#> RMSE                                |             |      |                |       |     |        |                    |   0.33

model <- glm(vs ~ disp, data = mtcars, family = "binomial")
r <- report(model)

#> Waiting for profiling to be done...

text_short(r)

#> We fitted a logistic model to predict vs with disp.The model's explanatory power is substantial (Tjur's R2 = 0.53). The model's intercept is at 4.14. Within this model:
#> 
#>   - The effect of (Intercept) is positive and can be considered as small and significant (beta = 4.14, 95% CI [1.81, 7.44], std. beta = -0.85, p < .01).
#>   - The effect of disp is negative and can be considered as large and significant (beta = -0.02, 95% CI [-0.04, -0.01], std. beta = -2.68, p < .01).

text_long(r)

#> We fitted a logistic model (estimated using ML) to predict vs with disp (formula = vs ~ disp). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Chen's (2010) recommendations.The model's explanatory power is substantial (Tjur's R2 = 0.53). The model's intercept, corresponding to vs = 0 and disp = 0, is at 4.14 (SE = 1.39, 95% CI [1.81, 7.44], p < .01). Within this model:
#> 
#>   - The effect of (Intercept) is positive and can be considered as small and significant (beta = 4.14, SE = 1.39, 95% CI [1.81, 7.44], std. beta = -0.85, p < .01).
#>   - The effect of disp is negative and can be considered as large and significant (beta = -0.02, SE = 0.01, 95% CI [-0.04, -0.01], std. beta = -2.68, p < .01).

table_short(r)

#> Parameter   | Coefficient |             CI |     p | Coefficient (std.) |  Fit
#> ------------------------------------------------------------------------------
#> (Intercept) |        4.14 | [ 1.81,  7.44] | 0.003 |              -0.85 |     
#> disp        |       -0.02 | [-0.04, -0.01] | 0.002 |              -2.68 |     
#>             |             |                |       |                    |     
#> Tjur's R2   |             |                |       |                    | 0.53

table_long(r)

#> Parameter       | Coefficient |   SE |             CI |     z |  df |     p | Coefficient (std.) |    Fit
#> ---------------------------------------------------------------------------------------------------------
#> (Intercept)     |        4.14 | 1.39 | [ 1.81,  7.44] |  2.98 | Inf | 0.003 |              -0.85 |       
#> disp            |       -0.02 | 0.01 | [-0.04, -0.01] | -3.03 | Inf | 0.002 |              -2.68 |       
#>                 |             |      |                |       |     |       |                    |       
#> AIC             |             |      |                |       |     |       |                    |  26.70
#> BIC             |             |      |                |       |     |       |                    |  29.63
#> Tjur's R2       |             |      |                |       |     |       |                    |   0.53
#> RMSE            |             |      |                |       |     |       |                    |   0.84
#> LOGLOSS         |             |      |                |       |     |       |                    |   0.35
#> SCORE_LOG       |             |      |                |       |     |       |                    | -11.07
#> SCORE_SPHERICAL |             |      |                |       |     |       |                    |   0.10
#> PCP             |             |      |                |       |     |       |                    |   0.77