(General) Linear Models Report

Create a report of a (general) linear model.

# S3 method for lm
report(model, ci = 0.95, standardize = TRUE,
  effsize = "cohen1988", performance_in_table = TRUE,
  performance_metrics = "all", bootstrap = FALSE, iterations = 500,
  ...)

# S3 method for glm
report(model, ci = 0.95, standardize = TRUE,
  effsize = "chen2010", performance_in_table = TRUE,
  performance_metrics = "all", bootstrap = FALSE, iterations = 500,
  ...)

Arguments

model
Object of class lm.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

standardize
Standardized coefficients. See model_parameters.

effsize
Interpret the standardized parameters using a set of rules. Can be "cohen1988" (default for linear models), "chen2010" (default for logistic models), "sawilowsky2009", NULL, or a custom set of rules.

performance_in_table
Add performance metrics in table.

performance_metrics
See model_performance.

bootstrap
See model_parameters.

iterations
The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.

...
Arguments passed to or from other methods.

Examples

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)

#> We fitted a linear model to predict Sepal.Length with Petal.Length and Species. The model's explanatory power is substantial (R2 = 0.84, adj. R2 = 0.83). The model's intercept is at 4.21. 
#> 
#> Within this model: 
#>   - Petal.Length is not significant (beta = 0.54, 95% CI [-0.00, 1.09], p = 0.05) and large (std. beta = 1.16).
#>   - Speciesversicolor is significant (beta = -1.81, 95% CI [-2.99, -0.62], p < .01) and large (std. beta = -0.88).
#>   - Speciesvirginica is significant (beta = -3.15, 95% CI [-4.41, -1.90], p < .001) and large (std. beta = -1.75).
#>   - Petal.Length:Speciesversicolor is not significant (beta = 0.29, 95% CI [-0.30, 0.87], p > .1) and medium (std. beta = 0.61).
#>   - Petal.Length:Speciesvirginica is not significant (beta = 0.45, 95% CI [-0.12, 1.03], p > .1) and large (std. beta = 0.97).

to_fulltext(r)

#> We fitted a linear model (using OLS algorithm) to predict Sepal.Length with Petal.Length and Species (formula = Sepal.Length ~ Petal.Length * Species). Effect sizes were labelled following Cohen's (1988) recommendations. The model explains a not significant and substantial proportion of variance (R2 = 0.84, F(, ) = , p , adj. R2 = 0.83). The model's intercept, corresponding to Petal.Length = 0 and Species = , is at 4.21 (t(144) = 10.34, 95% CI [3.41, 5.02], p < .001). 
#> 
#> Within this model: 
#>   - Petal.Length is positive, not significant (beta = 0.54, SE = 0.28, t(144) = 1.96, 95% CI [-0.00, 1.09], p = 0.05) and large (std. beta = 1.16, std. SE = 0.59, std. 95% CI [-0.01, 2.32]).
#>   - Speciesversicolor is negative, significant (beta = -1.81, SE = 0.60, t(144) = -3.02, 95% CI [-2.99, -0.62], p < .01) and large (std. beta = -0.88, std. SE = 0.77, std. 95% CI [-2.41, 0.65]).
#>   - Speciesvirginica is negative, significant (beta = -3.15, SE = 0.63, t(144) = -4.97, 95% CI [-4.41, -1.90], p < .001) and large (std. beta = -1.75, std. SE = 0.79, std. 95% CI [-3.32, -0.18]).
#>   - Petal.Length:Speciesversicolor is positive, not significant (beta = 0.29, SE = 0.30, t(144) = 0.97, 95% CI [-0.30, 0.87], p > .1) and medium (std. beta = 0.61, std. SE = 0.63, std. 95% CI [-0.63, 1.85]).
#>   - Petal.Length:Speciesvirginica is positive, not significant (beta = 0.45, SE = 0.29, t(144) = 1.56, 95% CI [-0.12, 1.03], p > .1) and large (std. beta = 0.97, std. SE = 0.62, std. 95% CI [-0.26, 2.19]).

to_table(r)

#> Parameter                      |  beta | CI_low | CI_high |    p | Std_beta |  Fit
#> ----------------------------------------------------------------------------------
#> (Intercept)                    |  4.21 |   3.41 |    5.02 | 0.00 |     0.49 |     
#> Petal.Length                   |  0.54 |  -0.00 |    1.09 | 0.05 |     1.16 |     
#> Speciesversicolor              | -1.81 |  -2.99 |   -0.62 | 0.00 |    -0.88 |     
#> Speciesvirginica               | -3.15 |  -4.41 |   -1.90 | 0.00 |    -1.75 |     
#> Petal.Length:Speciesversicolor |  0.29 |  -0.30 |    0.87 | 0.33 |     0.61 |     
#> Petal.Length:Speciesvirginica  |  0.45 |  -0.12 |    1.03 | 0.12 |     0.97 |     
#>                                |       |        |         |      |          |     
#> R2                             |       |        |         |      |          | 0.84
#> R2_adjusted                    |       |        |         |      |          | 0.83
#> 

to_fulltable(r)

#> Parameter                      |  beta |   SE | CI_low | CI_high |     t | DoF_residual |    p | Std_beta | Std_SE | Std_CI_low | Std_CI_high |    Fit
#> ------------------------------------------------------------------------------------------------------------------------------------------------------
#> (Intercept)                    |  4.21 | 0.41 |   3.41 |    5.02 | 10.34 |          144 | 0.00 |     0.49 |   0.77 |      -1.03 |        2.01 |       
#> Petal.Length                   |  0.54 | 0.28 |  -0.00 |    1.09 |  1.96 |          144 | 0.05 |     1.16 |   0.59 |      -0.01 |        2.32 |       
#> Speciesversicolor              | -1.81 | 0.60 |  -2.99 |   -0.62 | -3.02 |          144 | 0.00 |    -0.88 |   0.77 |      -2.41 |        0.65 |       
#> Speciesvirginica               | -3.15 | 0.63 |  -4.41 |   -1.90 | -4.97 |          144 | 0.00 |    -1.75 |   0.79 |      -3.32 |       -0.18 |       
#> Petal.Length:Speciesversicolor |  0.29 | 0.30 |  -0.30 |    0.87 |  0.97 |          144 | 0.33 |     0.61 |   0.63 |      -0.63 |        1.85 |       
#> Petal.Length:Speciesvirginica  |  0.45 | 0.29 |  -0.12 |    1.03 |  1.56 |          144 | 0.12 |     0.97 |   0.62 |      -0.26 |        2.19 |       
#>                                |       |      |        |         |       |              |      |          |        |            |             |       
#> AIC                            |       |      |        |         |       |              |      |          |        |            |             | 106.77
#> BIC                            |       |      |        |         |       |              |      |          |        |            |             | 127.84
#> R2                             |       |      |        |         |       |              |      |          |        |            |             |   0.84
#> R2_adjusted                    |       |      |        |         |       |              |      |          |        |            |             |   0.83
#> X0.329653168404125             |       |      |        |         |       |              |      |          |        |            |             |   0.33
#> 

model <- glm(vs ~ disp, data = mtcars, family = "binomial")
r <- report(model)
to_text(r)

#> We fitted a logistic model to predict vs with disp. The model's explanatory power is substantial (Tjur's R2 = 0.53). The model's intercept is at 4.14. 
#> 
#> Within this model: 
#>   - disp is significant (beta = -0.02, 95% CI [-0.04, -0.01], p < .01) and large (std. beta = -2.68).

to_fulltext(r)

#> We fitted a logistic model (using ML algorithm) to predict vs with disp (formula = vs ~ disp). Effect sizes were labelled following Chen's (2010) recommendations. The model's explanatory power is substantial (Tjur's R2 = 0.53). The model's intercept, corresponding to disp = 0, is at 4.14 (z = 2.98, 95% CI [1.81, 7.44], p < .01). 
#> 
#> Within this model: 
#>   - disp is negative, significant (beta = -0.02, SE = 0.01, z = -3.03, 95% CI [-0.04, -0.01], p < .01) and large (std. beta = -2.68, std. SE = 0.88, std. 95% CI [-4.90, -1.27]).

to_table(r)

#> Parameter   |  beta | CI_low | CI_high |    p | Std_beta |  Fit
#> ---------------------------------------------------------------
#> (Intercept) |  4.14 |   1.81 |    7.44 | 0.00 |    -0.85 |     
#> disp        | -0.02 |  -0.04 |   -0.01 | 0.00 |    -2.68 |     
#>             |       |        |         |      |          |     
#> R2_Tjur     |       |        |         |      |          | 0.53
#> 

to_fulltable(r)

#> Parameter   |  beta |   SE | CI_low | CI_high |     z | DoF_residual |    p | Std_beta | Std_SE | Std_CI_low | Std_CI_high |   Fit
#> ----------------------------------------------------------------------------------------------------------------------------------
#> (Intercept) |  4.14 | 1.39 |   1.81 |    7.44 |  2.98 |           30 | 0.00 |    -0.85 |   0.64 |      -2.42 |        0.27 |      
#> disp        | -0.02 | 0.01 |  -0.04 |   -0.01 | -3.03 |           30 | 0.00 |    -2.68 |   0.88 |      -4.90 |       -1.27 |      
#>             |       |      |        |         |       |              |      |          |        |            |             |      
#> AIC         |       |      |        |         |       |              |      |          |        |            |             | 26.70
#> BIC         |       |      |        |         |       |              |      |          |        |            |             | 29.63
#> R2_Tjur     |       |      |        |         |       |              |      |          |        |            |             |  0.53
#>