Linear Models Report

Create a report of a linear model.

# S3 method for lm
report(x, CI = 95, effsize = "cohen1988", data = NULL,
  ...)

Arguments

x
Object of class lm.

CI
Confidence Interval (CI) level. Default to 95%.

effsize
Compute standardized parameters and interpret them using a set of rules. Can be "cohen1988" (default), "sawilowsky2009", NULL, or a custom set of rules.

data
Dataframe used to fit the model. Can be useful to compute standardized parameters.

...
Arguments passed to or from other methods.

Examples

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)

#> We fitted a linear model to predict Sepal.Length with Petal.Length and Species. The model's explanatory power (R2) is of 0.84 (adj. R2 = 0.83). The model's intercept is at 4.21. 
#> 
#> Within this model: 
#>   - Petal.Length is not significant (beta = 0.54, 95% CI [0.00, 1.09], p = 0.05) and large (Std. beta = 1.16).
#>   - Speciesversicolor is significant (beta = -1.81, 95% CI [-2.99, -0.62], p < .01) and large (Std. beta = -0.88).
#>   - Speciesvirginica is significant (beta = -3.15, 95% CI [-4.41, -1.90], p < .001) and large (Std. beta = -1.75).
#>   - Petal.Length:Speciesversicolor is not significant (beta = 0.29, 95% CI [-0.30, 0.87], p > .1) and medium (Std. beta = 0.61).
#>   - Petal.Length:Speciesvirginica is not significant (beta = 0.45, 95% CI [-0.12, 1.03], p > .1) and large (Std. beta = 0.97).

to_fulltext(r)

#> We fitted a linear model to predict Sepal.Length with Petal.Length and Species (formula = Sepal.Length ~ Petal.Length * Species). Effect sizes were labelled following Cohen's (1988) recommendations. The model explains a significant proportion of variance (R2 = 0.84, F(6, 144) = 151.71, p < .001, adj. R2 = 0.83). The model's intercept is at 4.21(t() = 10.34, 95% CI [3.41, 5.02], p < .001). 
#> 
#> Within this model: 
#>   - Petal.Length is positive, not significant (beta = 0.54, t(144) = 1.96, 95% CI [0.00, 1.09], p = 0.05) and large (Std. beta = 1.16, Std. SE = 0.59, Std. 95% CI [-0.01, 2.32]).
#>   - Speciesversicolor is negative, significant (beta = -1.81, t(144) = -3.02, 95% CI [-2.99, -0.62], p < .01) and large (Std. beta = -0.88, Std. SE = 0.77, Std. 95% CI [-2.41, 0.65]).
#>   - Speciesvirginica is negative, significant (beta = -3.15, t(144) = -4.97, 95% CI [-4.41, -1.90], p < .001) and large (Std. beta = -1.75, Std. SE = 0.79, Std. 95% CI [-3.32, -0.18]).
#>   - Petal.Length:Speciesversicolor is positive, not significant (beta = 0.29, t(144) = 0.97, 95% CI [-0.30, 0.87], p > .1) and medium (Std. beta = 0.61, Std. SE = 0.63, Std. 95% CI [-0.63, 1.85]).
#>   - Petal.Length:Speciesvirginica is positive, not significant (beta = 0.45, t(144) = 1.56, 95% CI [-0.12, 1.03], p > .1) and large (Std. beta = 0.97, Std. SE = 0.62, Std. 95% CI [-0.26, 2.19]).

to_table(r)

#>                        Parameter       beta       CI_low    CI_high
#> 1                    (Intercept)  4.2131682  3.407870308  5.0184661
#> 2                   Petal.Length  0.5422926 -0.004757525  1.0893427
#> 3              Speciesversicolor -1.8056451 -2.988483690 -0.6228065
#> 4               Speciesvirginica -3.1535091 -4.406804100 -1.9002142
#> 5 Petal.Length:Speciesversicolor  0.2859884 -0.297224644  0.8692014
#> 6  Petal.Length:Speciesvirginica  0.4534460 -0.120048393  1.0269405
#>              p
#> 1 4.331619e-19
#> 2 5.199902e-02
#> 3 3.016413e-03
#> 4 1.846894e-06
#> 5 3.340471e-01
#> 6 1.202893e-01

to_fulltable(r)

#>                        Parameter       beta        SE         t DoF_residual
#> 1                    (Intercept)  4.2131682 0.4074209 10.341071          144
#> 2                   Petal.Length  0.5422926 0.2767667  1.959385          144
#> 3              Speciesversicolor -1.8056451 0.5984284 -3.017312          144
#> 4               Speciesvirginica -3.1535091 0.6340741 -4.973408          144
#> 5 Petal.Length:Speciesversicolor  0.2859884 0.2950624  0.969247          144
#> 6  Petal.Length:Speciesvirginica  0.4534460 0.2901455  1.562823          144
#>         CI_low    CI_high            p   Std_beta    Std_SE  Std_CI_low
#> 1  3.407870308  5.0184661 4.331619e-19  0.4924371 0.7695463 -1.02862887
#> 2 -0.004757525  1.0893427 5.199902e-02  1.1560769 0.5900202 -0.01014225
#> 3 -2.988483690 -0.6228065 3.016413e-03 -0.8826600 0.7741758 -2.41287658
#> 4 -4.406804100 -1.9002142 1.846894e-06 -1.7504145 0.7944200 -3.32064525
#> 5 -0.297224644  0.8692014 3.340471e-01  0.6096793 0.6290236 -0.63363314
#> 6 -0.120048393  1.0269405 1.202893e-01  0.9666710 0.6185417 -0.25592306
#>   Std_CI_high
#> 1   2.0135030
#> 2   2.3222961
#> 3   0.6475566
#> 4  -0.1801837
#> 5   1.8529917
#> 6   2.1892650