(General) Linear Models Report

Create a report of a (general) linear model.

# S3 method for lm
report(model, effsize = "default", ci = 0.95,
  standardize = "refit", standardize_robust = FALSE,
  bootstrap = FALSE, iterations = 500, performance_metrics = "all",
  ...)

Arguments

model
Frequentist regression object.

effsize
Interpret the standardized parameters using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

standardize
Add standardized parameters. Can be FALSE or a character indicating the standardization method (see parameters_standardize).

standardize_robust
Robust standardization. See parameters_standardize.

bootstrap
Should estimates be based on bootsrapped model? If TRUE, then arguments of Bayesian regressions apply.

iterations
The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.

performance_metrics
See model_performance.

...
Arguments passed to or from other methods (e.g., to standardize).

Examples

library(report)

model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
r <- report(model)
to_text(r)

#> We fitted a linear model to predict Sepal.Length with Petal.Length and Species. The model's explanatory power is substantial (R2 = 0.84, adj. R2 = 0.83). The model's intercept is at 4.21. Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and not significant (beta = 0.54, 95% CI [0.00, 1.09], p = 0.05, std. beta = 1.16).
#>   - The effect of Species (versicolor) is negative and can be considered as very large and significant (beta = -1.81, 95% CI [-2.99, -0.62], p < .01, std. beta = -0.88).
#>   - The effect of Species (virginica) is negative and can be considered as very large and significant (beta = -3.15, 95% CI [-4.41, -1.90], p < .001, std. beta = -1.75).
#>   - The interaction effect of Species (versicolor) on Petal.Length is positive and can be considered as medium and not significant (beta = 0.29, 95% CI [-0.30, 0.87], p > .1, std. beta = 0.61).
#>   - The interaction effect of Species (virginica) on Petal.Length is positive and can be considered as very large and not significant (beta = 0.45, 95% CI [-0.12, 1.03], p > .1, std. beta = 0.97).

to_fulltext(r)

#> We fitted a linear model (estimated using OLS) to predict Sepal.Length with Petal.Length and Species (formula = Sepal.Length ~ Petal.Length * Species). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Funder's (2019) recommendations.
#> 
#> 
#> 
#> The model explains a significant and substantial proportion of variance (R2 = 0.84, F(5, 144) = 151.71, p < .001, adj. R2 = 0.83). The model's intercept, corresponding to Sepal.Length = 0, Petal.Length = 0 and Species = setosa, is at 4.21 (SE = 0.41, 95% CI [3.41, 5.02], p < .001, std. intercept = 0.49). Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and not significant (beta = 0.54, SE = 0.28, 95% CI [0.00, 1.09], p = 0.05, std. beta = 1.16).
#>   - The effect of Species (versicolor) is negative and can be considered as very large and significant (beta = -1.81, SE = 0.60, 95% CI [-2.99, -0.62], p < .01, std. beta = -0.88).
#>   - The effect of Species (virginica) is negative and can be considered as very large and significant (beta = -3.15, SE = 0.63, 95% CI [-4.41, -1.90], p < .001, std. beta = -1.75).
#>   - The interaction effect of Species (versicolor) on Petal.Length is positive and can be considered as medium and not significant (beta = 0.29, SE = 0.30, 95% CI [-0.30, 0.87], p > .1, std. beta = 0.61).
#>   - The interaction effect of Species (virginica) on Petal.Length is positive and can be considered as very large and not significant (beta = 0.45, SE = 0.29, 95% CI [-0.12, 1.03], p > .1, std. beta = 0.97).

to_table(r)

#> Parameter                      | Coefficient |         95% CI |      p | Coefficient (std.) |  Fit
#> --------------------------------------------------------------------------------------------------
#> (Intercept)                    |        4.21 |   [3.41, 5.02] | < .001 |               0.49 |     
#> Petal.Length                   |        0.54 |   [0.00, 1.09] |   0.05 |               1.16 |     
#> Speciesversicolor              |       -1.81 | [-2.99, -0.62] |  < .01 |              -0.88 |     
#> Speciesvirginica               |       -3.15 | [-4.41, -1.90] | < .001 |              -1.75 |     
#> Petal.Length:Speciesversicolor |        0.29 |  [-0.30, 0.87] |   > .1 |               0.61 |     
#> Petal.Length:Speciesvirginica  |        0.45 |  [-0.12, 1.03] |   > .1 |               0.97 |     
#>                                |             |                |        |                    |     
#> R2                             |             |                |        |                    | 0.84
#> R2 (adj.)                      |             |                |        |                    | 0.83

to_fulltable(r)

#> Parameter                           | Coefficient |   SE |         95% CI |     t |  df |      p | Coefficient (std.) |    Fit
#> ------------------------------------------------------------------------------------------------------------------------------
#> (Intercept)                         |        4.21 | 0.41 |   [3.41, 5.02] | 10.34 | 144 | < .001 |               0.49 |       
#> Petal.Length                        |        0.54 | 0.28 |   [0.00, 1.09] |  1.96 | 144 |   0.05 |               1.16 |       
#> Species (versicolor)                |       -1.81 | 0.60 | [-2.99, -0.62] | -3.02 | 144 |  < .01 |              -0.88 |       
#> Species (virginica)                 |       -3.15 | 0.63 | [-4.41, -1.90] | -4.97 | 144 | < .001 |              -1.75 |       
#> Petal.Length * Species (versicolor) |        0.29 | 0.30 |  [-0.30, 0.87] |  0.97 | 144 |   > .1 |               0.61 |       
#> Petal.Length * Species (virginica)  |        0.45 | 0.29 |  [-0.12, 1.03] |  1.56 | 144 |   > .1 |               0.97 |       
#>                                     |             |      |                |       |     |        |                    |       
#> AIC                                 |             |      |                |       |     |        |                    | 106.77
#> BIC                                 |             |      |                |       |     |        |                    | 127.84
#> R2                                  |             |      |                |       |     |        |                    |   0.84
#> R2 (adj.)                           |             |      |                |       |     |        |                    |   0.83
#> RMSE                                |             |      |                |       |     |        |                    |   0.33

model <- glm(vs ~ disp, data = mtcars, family = "binomial")
r <- report(model)
to_text(r)

#> We fitted a logistic model to predict vs with disp. The model's explanatory power is substantial (Tjur's R2 = 0.53). The model's intercept is at 4.14. Within this model:
#> 
#>   - The effect of disp is negative and can be considered as large and significant (beta = -0.02, 95% CI [-0.04, -0.01], p < .01, std. beta = -2.68).

to_fulltext(r)

#> We fitted a logistic model (estimated using ML) to predict vs with disp (formula = vs ~ disp). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Chen's (2010) recommendations.
#> 
#> The model's explanatory power is substantial (Tjur's R2 = 0.53). The model's intercept, corresponding to vs = 0 and disp = 0, is at 4.14 (SE = 1.39, 95% CI [1.81, 7.44], p < .01, std. intercept = -0.85). Within this model:
#> 
#>   - The effect of disp is negative and can be considered as large and significant (beta = -0.02, SE = 0.01, 95% CI [-0.04, -0.01], p < .01, std. beta = -2.68).

to_table(r)

#> Parameter   | Coefficient |         95% CI |     p | Coefficient (std.) |  Fit
#> ------------------------------------------------------------------------------
#> (Intercept) |        4.14 |   [1.81, 7.44] | < .01 |              -0.85 |     
#> disp        |       -0.02 | [-0.04, -0.01] | < .01 |              -2.68 |     
#>             |             |                |       |                    |     
#> Tjur's R2   |             |                |       |                    | 0.53

to_fulltable(r)

#> Parameter       | Coefficient |   SE |         95% CI |     z | df |     p | Coefficient (std.) |    Fit
#> --------------------------------------------------------------------------------------------------------
#> (Intercept)     |        4.14 | 1.39 |   [1.81, 7.44] |  2.98 | 30 | < .01 |              -0.85 |       
#> disp            |       -0.02 | 0.01 | [-0.04, -0.01] | -3.03 | 30 | < .01 |              -2.68 |       
#>                 |             |      |                |       |    |       |                    |       
#> AIC             |             |      |                |       |    |       |                    |  26.70
#> BIC             |             |      |                |       |    |       |                    |  29.63
#> Tjur's R2       |             |      |                |       |    |       |                    |   0.53
#> RMSE            |             |      |                |       |    |       |                    |   0.84
#> LOGLOSS         |             |      |                |       |    |       |                    |   0.35
#> SCORE_LOG       |             |      |                |       |    |       |                    | -11.07
#> SCORE_SPHERICAL |             |      |                |       |    |       |                    |   0.10
#> PCP             |             |      |                |       |    |       |                    |   0.77