Mixed Models Report

Create a report of a mixed model (lme4).

# S3 method for lmerMod
report(
  model,
  interpretation = "default",
  ci = 0.95,
  standardize = "refit",
  standardize_robust = FALSE,
  bootstrap = FALSE,
  iterations = 500,
  performance_metrics = "all",
  df_method = "wald",
  ...
)

Arguments

model
Model object.

interpretation
Interpret the standardized parameters using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

standardize
The method used for standardizing the parameters. Can be "refit", "posthoc", "smart", "basic" or NULL (default) for no standardization. See 'Details' in standardize_parameters. Note that robust estimation (i.e. robust=TRUE) of standardized parameters only works when standardize="refit".

standardize_robust
Logical, if TRUE, robust standard errors are calculated (if possible), and confidence intervals and p-values are based on these robust standard errors.

bootstrap
Should estimates be based on bootstrapped model? If TRUE, then arguments of Bayesian regressions apply (see also bootstrap_parameters()).

iterations
The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.

performance_metrics
See model_performance.

df_method
Method for computing degrees of freedom for p values, standard errors and confidence intervals (CI). May be "wald" (default, see degrees_of_freedom), "ml1" (see dof_ml1), "betwithin" (see dof_betwithin), "satterthwaite" (see dof_satterthwaite) or "kenward" (see dof_kenward). Note that when df_method is not "wald", robust standard errors etc. cannot be computed.

...
Arguments passed to or from other methods. For instance, when bootstrap = TRUE, arguments like ci_method are passed down to describe_posterior.

Examples

library(lme4)
library(report)

model <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
r <- report(model)
text_short(r)

#> We fitted a linear mixed model to predict Sepal.Length with Petal.Length. The model included Species as random effects.The model's total explanatory power is substantial (conditional R2 = 0.97) and the part related to the fixed effects alone (marginal R2) is of 0.66. The model's intercept is at 2.50. Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and significant (beta = 0.89, std. beta = 1.89, p < .001).

text_long(r)

#> We fitted a linear mixed model (estimated using REML and nloptwrap optimizer) to predict Sepal.Length with Petal.Length (formula = Sepal.Length ~ Petal.Length). The model included Species as random effects (formula = ~1 | Species). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Funder's (2019) recommendations.The model's total explanatory power is substantial (conditional R2 = 0.97) and the part related to the fixed effects alone (marginal R2) is of 0.66. The model's intercept, corresponding to Sepal.Length = 0, Petal.Length = 0 and Species = setosa, is at 2.50 (SE = 0.67, 95% CI [1.20, 3.81], p < .001). Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and significant (beta = 0.89, SE = 0.06, std. beta = 1.89, p < .001).

table_short(r)

#> Parameter        | Coefficient |        p | Std_Coefficient |  Fit
#> ------------------------------------------------------------------
#> (Intercept)      |        2.50 | 1.75e-04 |       -1.46e-13 |     
#> Petal.Length     |        0.89 | 4.34e-44 |            1.89 |     
#>                  |             |          |                 |     
#> R2 (conditional) |             |          |                 | 0.97
#> R2 (marginal)    |             |          |                 | 0.66

table_long(r)

#> Parameter        | Coefficient |   SE | CI_low.x | CI_high.x |     t | df_error |        p | Std_Coefficient |   CI | CI_low.y | CI_high.y |    Fit
#> ---------------------------------------------------------------------------------------------------------------------------------------------------
#> (Intercept)      |        2.50 | 0.67 |     1.20 |      3.81 |  3.75 |      146 | 1.75e-04 |       -1.46e-13 | 0.95 |    -1.47 |      1.47 |       
#> Petal.Length     |        0.89 | 0.06 |     0.76 |      1.01 | 13.93 |      146 | 4.34e-44 |            1.89 | 0.95 |     1.63 |      2.16 |       
#>                  |             |      |          |           |       |          |          |                 |      |          |           |       
#> AIC              |             |      |          |           |       |          |          |                 |      |          |           | 127.79
#> BIC              |             |      |          |           |       |          |          |                 |      |          |           | 139.84
#> R2 (conditional) |             |      |          |           |       |          |          |                 |      |          |           |   0.97
#> R2 (marginal)    |             |      |          |           |       |          |          |                 |      |          |           |   0.66
#> ICC              |             |      |          |           |       |          |          |                 |      |          |           |   0.91
#> RMSE             |             |      |          |           |       |          |          |                 |      |          |           |   0.33

model <- glmer(vs ~ disp + (1 | am), data = mtcars, family = "binomial")
r <- report(model)
text_short(r)

#> We fitted a logistic mixed model to predict vs with disp. The model included am as random effects.The model's total explanatory power is substantial (conditional R2 = 0.99) and the part related to the fixed effects alone (marginal R2) is of 0.77. The model's intercept is at 25.49. Within this model:
#> 
#>   - The effect of disp is negative and can be considered as large and not significant (beta = -0.13, std. beta = -16.53, p = 0.236).

text_long(r)

#> We fitted a logistic mixed model (estimated using ML and Nelder-Mead optimizer) to predict vs with disp (formula = vs ~ disp). The model included am as random effects (formula = ~1 | am). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Chen's (2010) recommendations.The model's total explanatory power is substantial (conditional R2 = 0.99) and the part related to the fixed effects alone (marginal R2) is of 0.77. The model's intercept, corresponding to vs = 0, disp = 0 and am = 0, is at 25.49 (SE = 22.84, 95% CI [-19.27, 70.24], p = 0.264). Within this model:
#> 
#>   - The effect of disp is negative and can be considered as large and not significant (beta = -0.13, SE = 0.11, std. beta = -16.53, p = 0.236).

table_short(r)

#> Parameter        | Coefficient |    p | Std_Coefficient |  Fit
#> --------------------------------------------------------------
#> (Intercept)      |       25.49 | 0.26 |           -5.28 |     
#> disp             |       -0.13 | 0.24 |          -16.53 |     
#>                  |             |      |                 |     
#> R2 (conditional) |             |      |                 | 0.99
#> R2 (marginal)    |             |      |                 | 0.77

table_long(r)

#> Parameter        | Coefficient |    SE | CI_low.x | CI_high.x |     z | df_error |    p | Std_Coefficient |   CI | CI_low.y | CI_high.y |    Fit
#> ------------------------------------------------------------------------------------------------------------------------------------------------
#> (Intercept)      |       25.49 | 22.84 |   -19.27 |     70.24 |  1.12 |      Inf | 0.26 |           -5.28 | 0.95 |   -20.17 |      9.62 |       
#> disp             |       -0.13 |  0.11 |    -0.35 |      0.09 | -1.18 |      Inf | 0.24 |          -16.53 | 0.95 |   -43.88 |     10.83 |       
#>                  |             |       |          |           |       |          |      |                 |      |          |           |       
#> AIC              |             |       |          |           |       |          |      |                 |      |          |           |  21.12
#> BIC              |             |       |          |           |       |          |      |                 |      |          |           |  25.51
#> R2 (conditional) |             |       |          |           |       |          |      |                 |      |          |           |   0.99
#> R2 (marginal)    |             |       |          |           |       |          |      |                 |      |          |           |   0.77
#> ICC              |             |       |          |           |       |          |      |                 |      |          |           |   0.96
#> RMSE             |             |       |          |           |       |          |      |                 |      |          |           |   0.40
#> LOGLOSS          |             |       |          |           |       |          |      |                 |      |          |           |   0.08
#> SCORE_LOG        |             |       |          |           |       |          |      |                 |      |          |           | -44.95
#> SCORE_SPHERICAL  |             |       |          |           |       |          |      |                 |      |          |           |   0.12