Mixed Models Report

Create a report of a mixed model (lme4).

# S3 method for lmerMod
report(model, effsize = "default", ci = 0.95,
  standardize = "refit", standardize_robust = FALSE,
  bootstrap = FALSE, iterations = 500, performance_metrics = "all",
  p_method = "wald", ci_method = "wald", ...)

Arguments

model
Frequentist regression object.

effsize
Interpret the standardized parameters using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

standardize
Add standardized parameters. Can be FALSE or a character indicating the standardization method (see parameters_standardize).

standardize_robust
Robust standardization. See parameters_standardize.

bootstrap
Should estimates be based on bootsrapped model? If TRUE, then arguments of Bayesian regressions apply.

iterations
The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.

performance_metrics
See model_performance.

p_method
Method for computing p values. See p_value.

ci_method
Method for computing confidence intervals (CI). See ci.

...
Arguments passed to or from other methods (e.g., to standardize).

Examples

library(lme4)

#> Loading required package: Matrix

library(report)

model <- lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
r <- report(model)
to_text(r)

#> We fitted a linear mixed model to predict Sepal.Length with Petal.Length. The model included Species as random effects. The model's total explanatory power is substantial (conditional R2 = 0.97) and the part related to the fixed effects alone (marginal R2) is of 0.66. The model's intercept is at 2.50. Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and significant (beta = 0.89, 95% CI [-0.85, 2.63], p < .001, std. beta = 1.89).

to_fulltext(r)

#> We fitted a linear mixed model (estimated using REML and nloptwrap optimizer) to predict Sepal.Length with Petal.Length (formula = Sepal.Length ~ Petal.Length). The model included Species as random effects (formula = ~1 | Species). The 95% Confidence Intervals (CIs) and p values were computed using Wald approximation and . Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Funder's (2019) recommendations.
#> 
#> The model's total explanatory power is substantial (conditional R2 = 0.97) and the part related to the fixed effects alone (marginal R2) is of 0.66. The model's intercept, corresponding to Sepal.Length = 0, Petal.Length = 0 and Species = setosa, is at 2.50 (SE = 0.67, 95% CI [-2.40, 7.41], p < .001, std. intercept = 0.00). Within this model:
#> 
#>   - The effect of Petal.Length is positive and can be considered as very large and significant (beta = 0.89, SE = 0.06, 95% CI [-0.85, 2.63], p < .001, std. beta = 1.89).

to_table(r)

#> Parameter      | Coefficient |        95% CI |      p | Coefficient (std.) |  Fit
#> ---------------------------------------------------------------------------------
#> (Intercept)    |        2.50 | [-2.40, 7.41] | < .001 |               0.00 |     
#> Petal.Length   |        0.89 | [-0.85, 2.63] | < .001 |               1.89 |     
#>                |             |               |        |                    |     
#> R2_conditional |             |               |        |                    | 0.97
#> R2_marginal    |             |               |        |                    | 0.66

to_fulltable(r)

#> Parameter      | Coefficient |   SE |        95% CI |     t |      p | Coefficient (std.) |    Fit
#> --------------------------------------------------------------------------------------------------
#> (Intercept)    |        2.50 | 0.67 | [-2.40, 7.41] |  3.75 | < .001 |               0.00 |       
#> Petal.Length   |        0.89 | 0.06 | [-0.85, 2.63] | 13.93 | < .001 |               1.89 |       
#>                |             |      |               |       |        |                    |       
#> AIC            |             |      |               |       |        |                    | 127.79
#> BIC            |             |      |               |       |        |                    | 139.84
#> R2_conditional |             |      |               |       |        |                    |   0.97
#> R2_marginal    |             |      |               |       |        |                    |   0.66
#> ICC            |             |      |               |       |        |                    |   0.91
#> RMSE           |             |      |               |       |        |                    |   0.33

model <- glmer(vs ~ disp + (1 | am), data = mtcars, family = "binomial")
r <- report(model)

#> Warning: Variable `c(1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, ` contains only two different values. Consider converting it to a factor.Variable `0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1)` contains only two different values. Consider converting it to a factor.

#> Warning: Variable `c(1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, ` contains only two different values. Consider converting it to a factor.Variable `0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1)` contains only two different values. Consider converting it to a factor.

#> Warning: Variable `c(1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, ` contains only two different values. Consider converting it to a factor.Variable `0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1)` contains only two different values. Consider converting it to a factor.

#> Warning: Variable `c(1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, ` contains only two different values. Consider converting it to a factor.Variable `0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1)` contains only two different values. Consider converting it to a factor.

to_text(r)

#> We fitted a logistic mixed model to predict vs with disp. The model included am as random effects. The model's total explanatory power is substantial (conditional R2 = 0.99) and the part related to the fixed effects alone (marginal R2) is of 0.77. The model's intercept is at 25.49. Within this model:
#> 
#>   - The effect of disp is negative and can be considered as large and not significant (beta = -0.13, 95% CI [0.13, -0.39], p > .1, std. beta = -16.53).

to_fulltext(r)

#> We fitted a logistic mixed model (estimated using ML and Nelder-Mead optimizer) to predict vs with disp (formula = vs ~ disp). The model included am as random effects (formula = ~1 | am). The 95% Confidence Intervals (CIs) and p values were computed using Wald approximation and . Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Chen's (2010) recommendations.
#> 
#> The model's total explanatory power is substantial (conditional R2 = 0.99) and the part related to the fixed effects alone (marginal R2) is of 0.77. The model's intercept, corresponding to vs = 0, disp = 0 and am = 0, is at 25.49 (SE = 22.83, 95% CI [-24.47, 75.44], p > .1, std. intercept = -5.28). Within this model:
#> 
#>   - The effect of disp is negative and can be considered as large and not significant (beta = -0.13, SE = 0.11, 95% CI [0.13, -0.39], p > .1, std. beta = -16.53).

to_table(r)

#> Parameter      | Coefficient |          95% CI |    p | Coefficient (std.) |  Fit
#> ---------------------------------------------------------------------------------
#> (Intercept)    |       25.49 | [-24.47, 75.44] | > .1 |              -5.28 |     
#> disp           |       -0.13 |   [0.13, -0.39] | > .1 |             -16.53 |     
#>                |             |                 |      |                    |     
#> R2_conditional |             |                 |      |                    | 0.99
#> R2_marginal    |             |                 |      |                    | 0.77

to_fulltable(r)

#> Parameter       | Coefficient |    SE |          95% CI |     z |    p | Coefficient (std.) |    Fit
#> ----------------------------------------------------------------------------------------------------
#> (Intercept)     |       25.49 | 22.83 | [-24.47, 75.44] |  1.12 | > .1 |              -5.28 |       
#> disp            |       -0.13 |  0.11 |   [0.13, -0.39] | -1.18 | > .1 |             -16.53 |       
#>                 |             |       |                 |       |      |                    |       
#> AIC             |             |       |                 |       |      |                    |  21.12
#> BIC             |             |       |                 |       |      |                    |  25.51
#> R2_conditional  |             |       |                 |       |      |                    |   0.99
#> R2_marginal     |             |       |                 |       |      |                    |   0.77
#> ICC             |             |       |                 |       |      |                    |   0.96
#> RMSE            |             |       |                 |       |      |                    |   0.40
#> LOGLOSS         |             |       |                 |       |      |                    |   0.08
#> SCORE_LOG       |             |       |                 |       |      |                    | -44.95
#> SCORE_SPHERICAL |             |       |                 |       |      |                    |   0.12