Report R environment (packages, system, etc.)

Report R environment (packages, system, etc.)

# S3 method for sessionInfo
report(x, ...)

report_packages(session = NULL, include_R = TRUE, ...)

cite_packages(session = NULL, include_R = TRUE, ...)

report_system(session = NULL)

Arguments

x
The R object that you want to report (see list of of supported objects above).

...
Arguments passed to or from other methods.

session
A sessionInfo object.

include_R
Include R in the citations.

Value

For report_packages, a data frame of class with information on package name, version and citation.

An object of class report().

Examples

library(report)

session <- sessionInfo()

r <- report(session)
r
#> Analyses were conducted using the R Statistical language (version 4.1.1; R Core Team, 2021) on macOS Catalina 10.15.7, using the packages Matrix (version 1.3.4; Douglas Bates and Martin Maechler, 2021), lme4 (version 1.1.27.1; Douglas Bates et al., 2015), dplyr (version 1.0.7; Hadley Wickham et al., 2021), performance (version 0.7.3.5; Lüdecke et al., 2021), bayestestR (version 0.11.0; Makowski et al., 2019), report (version 0.4.0; Makowski et al., 2020) and lavaan (version 0.6.9; Yves Rosseel, 2012).
#> 
#> References
#> ----------
#>   - Douglas Bates and Martin Maechler (2021). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.3-4. https://CRAN.R-project.org/package=Matrix
#>   - Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#>   - Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2021). dplyr: A Grammar of Data Manipulation. R package version 1.0.7. https://CRAN.R-project.org/package=dplyr
#>   - Lüdecke et al., (2021). performance: An R Package for Assessment, Comparison and Testing of Statistical Models. Journal of Open Source Software, 6(60), 3139. https://doi.org/10.21105/joss.03139
#>   - Makowski, D., Ben-Shachar, M., & Lüdecke, D. (2019). bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework. Journal of Open Source Software, 4(40), 1541. doi:10.21105/joss.01541
#>   - Makowski, D., Ben-Shachar, M.S., Patil, I. & Lüdecke, D. (2020). Automated Results Reporting as a Practical Tool to Improve Reproducibility and Methodological Best Practices Adoption. CRAN. Available from https://github.com/easystats/report. doi: .
#>   - R Core Team (2021). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
#>   - Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL https://www.jstatsoft.org/v48/i02/.
summary(r)
#> The analysis was done using the R Statistical language (v4.1.1; R Core Team, 2021) on macOS Catalina 10.15.7, using the packages Matrix (v1.3.4), lme4 (v1.1.27.1), dplyr (v1.0.7), performance (v0.7.3.5), bayestestR (v0.11.0), report (v0.4.0) and lavaan (v0.6.9).
as.data.frame(r)
#> Package     |  Version |                                                                                                                                                                                                                                                Reference
#> ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> Matrix      |    1.3.4 |                                                                                        Douglas Bates and Martin Maechler (2021). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.3-4. https://CRAN.R-project.org/package=Matrix
#> R           |    4.1.1 |                                                                                    R Core Team (2021). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
#> bayestestR  |   0.11.0 |               Makowski, D., Ben-Shachar, M., & Lüdecke, D. (2019). bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework. Journal of Open Source Software, 4(40), 1541. doi:10.21105/joss.01541
#> dplyr       |    1.0.7 |                                                                         Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2021). dplyr: A Grammar of Data Manipulation. R package version 1.0.7. https://CRAN.R-project.org/package=dplyr
#> lavaan      |    0.6.9 |                                                                                        Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL https://www.jstatsoft.org/v48/i02/.
#> lme4        | 1.1.27.1 |                                                                Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#> performance |  0.7.3.5 |                                                        Lüdecke et al., (2021). performance: An R Package for Assessment, Comparison and Testing of Statistical Models. Journal of Open Source Software, 6(60), 3139. https://doi.org/10.21105/joss.03139
#> report      |    0.4.0 | Makowski, D., Ben-Shachar, M.S., Patil, I. & Lüdecke, D. (2020). Automated Results Reporting as a Practical Tool to Improve Reproducibility and Methodological Best Practices Adoption. CRAN. Available from https://github.com/easystats/report. doi: .
summary(as.data.frame(r))
#> Package     |  Version
#> ----------------------
#> Matrix      |    1.3.4
#> R           |    4.1.1
#> bayestestR  |   0.11.0
#> dplyr       |    1.0.7
#> lavaan      |    0.6.9
#> lme4        | 1.1.27.1
#> performance |  0.7.3.5
#> report      |    0.4.0

# Convenience functions
report_packages(include_R = FALSE)
#>   - Matrix (version 1.3.4; Douglas Bates and Martin Maechler, 2021)
#>   - lme4 (version 1.1.27.1; Douglas Bates et al., 2015)
#>   - dplyr (version 1.0.7; Hadley Wickham et al., 2021)
#>   - performance (version 0.7.3.5; Lüdecke et al., 2021)
#>   - bayestestR (version 0.11.0; Makowski et al., 2019)
#>   - report (version 0.4.0; Makowski et al., 2020)
#>   - lavaan (version 0.6.9; Yves Rosseel, 2012)
cite_packages(prefix = "> ")
#> > Douglas Bates and Martin Maechler (2021). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.3-4. https://CRAN.R-project.org/package=Matrix
#> > Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#> > Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2021). dplyr: A Grammar of Data Manipulation. R package version 1.0.7. https://CRAN.R-project.org/package=dplyr
#> > Lüdecke et al., (2021). performance: An R Package for Assessment, Comparison and Testing of Statistical Models. Journal of Open Source Software, 6(60), 3139. https://doi.org/10.21105/joss.03139
#> > Makowski, D., Ben-Shachar, M., & Lüdecke, D. (2019). bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework. Journal of Open Source Software, 4(40), 1541. doi:10.21105/joss.01541
#> > Makowski, D., Ben-Shachar, M.S., Patil, I. & Lüdecke, D. (2020). Automated Results Reporting as a Practical Tool to Improve Reproducibility and Methodological Best Practices Adoption. CRAN. Available from https://github.com/easystats/report. doi: .
#> > R Core Team (2021). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
#> > Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL https://www.jstatsoft.org/v48/i02/.
report_system()
#> Analyses were conducted using the R Statistical language (version 4.1.1; R Core Team, 2021) on macOS Catalina 10.15.7