Report R environment (packages, system, etc.)

Report R environment (packages, system, etc.)

# S3 method for sessionInfo
report(x, ...)

report_packages(session = NULL, include_R = TRUE, ...)

cite_packages(session = NULL, include_R = TRUE, ...)

report_system(session = NULL)

Arguments

x
The R object that you want to report (see list of of supported objects above).

...
Arguments passed to or from other methods.

session
A sessionInfo object.

include_R
Include R in the citations.

Value

For report_packages, a data frame of class with information on package name, version and citation.

Examples

library(report)

session <- sessionInfo()

r <- report(session)
r

#> Analyses were conducted using the R Statistical language (version 4.0.3; R Core Team, 2020) on macOS Catalina 10.15.7, using the packages Matrix (version 1.2.18; Douglas Bates and Martin Maechler, 2019), lme4 (version 1.1.26; Douglas Bates et al., 2015), dplyr (version 1.0.2; Hadley Wickham et al., 2020), bayestestR (version 0.8.0.1; Makowski et al., 2019), report (version 0.2.0; Makowski et al., 2020) and lavaan (version 0.6.7; Yves Rosseel, 2012).
#> 
#> References
#> ----------
#>   - Douglas Bates and Martin Maechler (2019). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.2-18. https://CRAN.R-project.org/package=Matrix
#>   - Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#>   - Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2020). dplyr: A Grammar of Data Manipulation. R package version 1.0.2. https://CRAN.R-project.org/package=dplyr
#>   - Makowski, D., Ben-Shachar, M., & Lüdecke, D. (2019). bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework. Journal of Open Source Software, 4(40), 1541. doi:10.21105/joss.01541
#>   - Makowski, D., Lüdecke, D., & Ben-Shachar, M.S. (2020). Automated reporting as a practical tool to improve reproducibility and methodological best practices adoption. CRAN. Available from https://github.com/easystats/report. doi: .
#>   - R Core Team (2020). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
#>   - Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL http://www.jstatsoft.org/v48/i02/.

summary(r)

#> The analysis was done using the R Statistical language (v4.0.3; R Core Team, 2020) on macOS Catalina 10.15.7, using the packages Matrix (v1.2.18), lme4 (v1.1.26), dplyr (v1.0.2), bayestestR (v0.8.0.1), report (v0.2.0) and lavaan (v0.6.7).

as.data.frame(r)

#> Package    | Version |                                                                                                                                                                                                                                  Reference
#> -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> bayestestR | 0.8.0.1 | Makowski, D., Ben-Shachar, M., & Lüdecke, D. (2019). bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework. Journal of Open Source Software, 4(40), 1541. doi:10.21105/joss.01541
#> dplyr      |   1.0.2 |                                                           Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2020). dplyr: A Grammar of Data Manipulation. R package version 1.0.2. https://CRAN.R-project.org/package=dplyr
#> lavaan     |   0.6.7 |                                                                           Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL http://www.jstatsoft.org/v48/i02/.
#> lme4       |  1.1.26 |                                                  Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#> Matrix     |  1.2.18 |                                                                         Douglas Bates and Martin Maechler (2019). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.2-18. https://CRAN.R-project.org/package=Matrix
#> R          |   4.0.3 |                                                                      R Core Team (2020). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
#> report     |   0.2.0 |     Makowski, D., Lüdecke, D., & Ben-Shachar, M.S. (2020). Automated reporting as a practical tool to improve reproducibility and methodological best practices adoption. CRAN. Available from https://github.com/easystats/report. doi: .

summary(as.data.frame(r))

#> Package    | Version
#> --------------------
#> bayestestR | 0.8.0.1
#> dplyr      |   1.0.2
#> lavaan     |   0.6.7
#> lme4       |  1.1.26
#> Matrix     |  1.2.18
#> R          |   4.0.3
#> report     |   0.2.0

# Convenience functions
report_packages(include_R = FALSE)

#>   - Matrix (version 1.2.18; Douglas Bates and Martin Maechler, 2019)
#>   - lme4 (version 1.1.26; Douglas Bates et al., 2015)
#>   - dplyr (version 1.0.2; Hadley Wickham et al., 2020)
#>   - bayestestR (version 0.8.0.1; Makowski et al., 2019)
#>   - report (version 0.2.0; Makowski et al., 2020)
#>   - lavaan (version 0.6.7; Yves Rosseel, 2012)

cite_packages(prefix = "> ")

#> > Douglas Bates and Martin Maechler (2019). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.2-18. https://CRAN.R-project.org/package=Matrix
#> > Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#> > Hadley Wickham, Romain François, Lionel Henry and Kirill Müller (2020). dplyr: A Grammar of Data Manipulation. R package version 1.0.2. https://CRAN.R-project.org/package=dplyr
#> > Makowski, D., Ben-Shachar, M., & Lüdecke, D. (2019). bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework. Journal of Open Source Software, 4(40), 1541. doi:10.21105/joss.01541
#> > Makowski, D., Lüdecke, D., & Ben-Shachar, M.S. (2020). Automated reporting as a practical tool to improve reproducibility and methodological best practices adoption. CRAN. Available from https://github.com/easystats/report. doi: .
#> > R Core Team (2020). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
#> > Yves Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36. URL http://www.jstatsoft.org/v48/i02/.

report_system()

#> Analyses were conducted using the R Statistical language (version 4.0.3; R Core Team, 2020) on macOS Catalina 10.15.7