Report R environment (packages, system, etc.)

Report R environment (packages, system, etc.)

Usage

# S3 method for sessionInfo
report(x, ...)

report_packages(session = NULL, include_R = TRUE, ...)

cite_packages(session = NULL, include_R = TRUE, ...)

report_system(session = NULL)

Arguments

x

The R object that you want to report (see list of of supported objects above).

...

Arguments passed to or from other methods.

session

A sessionInfo object.

include_R

Include R in the citations.

Value

For report_packages, a data frame of class with information on package name, version and citation.

An object of class report().

Examples

library(report)

session <- sessionInfo()

r <- report(session)
r
#> Analyses were conducted using the R Statistical language (version 4.2.2; R Core
#> Team, 2022) on Ubuntu 22.04.1 LTS, using the packages Matrix (version 1.5.3;
#> Bates D et al., 2022), lme4 (version 1.1.31; Bates D et al., 2015), performance
#> (version 0.10.2; Lüdecke D et al., 2021), bayestestR (version 0.13.0; Makowski
#> D et al., 2019), report (version 0.5.6; Makowski D et al., 2023), lavaan
#> (version 0.6.13; Rosseel Y, 2012) and dplyr (version 1.1.0; Wickham H et al.,
#> 2023).
#> 
#> References
#> ----------
#>   - Bates D, Maechler M, Jagan M (2022). _Matrix: Sparse and Dense Matrix Classes
#> and Methods_. https://Matrix.R-forge.R-project.org/,
#> https://Matrix.R-forge.R-project.org/doxygen/.
#>   - Bates D, Mächler M, Bolker B, Walker S (2015). “Fitting Linear Mixed-Effects
#> Models Using lme4.” _Journal of Statistical Software_, *67*(1), 1-48.
#> doi:10.18637/jss.v067.i01 <https://doi.org/10.18637/jss.v067.i01>.
#>   - Lüdecke D, Ben-Shachar M, Patil I, Waggoner P, Makowski D (2021).
#> “performance: An R Package for Assessment, Comparison and Testing of
#> Statistical Models.” _Journal of Open Source Software_, *6*(60), 3139.
#> doi:10.21105/joss.03139 <https://doi.org/10.21105/joss.03139>.
#>   - Makowski D, Ben-Shachar M, Lüdecke D (2019). “bayestestR: Describing Effects
#> and their Uncertainty, Existence and Significance within the Bayesian
#> Framework.” _Journal of Open Source Software_, *4*(40), 1541.
#> doi:10.21105/joss.01541 <https://doi.org/10.21105/joss.01541>,
#> <https://joss.theoj.org/papers/10.21105/joss.01541>.
#>   - Makowski D, Lüdecke D, Patil I, Thériault R (2023). “Automated Results
#> Reporting as a Practical Tool to Improve Reproducibility and Methodological
#> Best Practices Adoption.” _CRAN_. <https://easystats.github.io/report/>.
#>   - R Core Team (2022). _R: A Language and Environment for Statistical
#> Computing_. R Foundation for Statistical Computing, Vienna, Austria.
#> <https://www.R-project.org/>.
#>   - Rosseel Y (2012). “lavaan: An R Package for Structural Equation Modeling.”
#> _Journal of Statistical Software_, *48*(2), 1-36. doi:10.18637/jss.v048.i02
#> <https://doi.org/10.18637/jss.v048.i02>.
#>   - Wickham H, François R, Henry L, Müller K, Vaughan D (2023). _dplyr: A Grammar
#> of Data Manipulation_. https://dplyr.tidyverse.org,
#> https://github.com/tidyverse/dplyr.
summary(r)
#> The analysis was done using the R Statistical language (v4.2.2; R Core Team,
#> 2022) on Ubuntu 22.04.1 LTS, using the packages Matrix (v1.5.3), lme4
#> (v1.1.31), performance (v0.10.2), bayestestR (v0.13.0), report (v0.5.6), lavaan
#> (v0.6.13) and dplyr (v1.1.0).
as.data.frame(r)
#> Package     | Version |                                                                                                                                                                                                                                                                                                                            Reference
#> ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
#> Matrix      |   1.5.3 |                                                                                                                                                    Bates D, Maechler M, Jagan M (2022). _Matrix: Sparse and Dense Matrix Classes and Methods_. https://Matrix.R-forge.R-project.org/, https://Matrix.R-forge.R-project.org/doxygen/.
#> R           |   4.2.2 |                                                                                                                                                                R Core Team (2022). _R: A Language and Environment for Statistical Computing_. R Foundation for Statistical Computing, Vienna, Austria. <https://www.R-project.org/>.
#> bayestestR  |  0.13.0 | Makowski D, Ben-Shachar M, Lüdecke D (2019). “bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework.” _Journal of Open Source Software_, *4*(40), 1541. doi:10.21105/joss.01541 <https://doi.org/10.21105/joss.01541>, <https://joss.theoj.org/papers/10.21105/joss.01541>.
#> dplyr       |   1.1.0 |                                                                                                                                                                Wickham H, François R, Henry L, Müller K, Vaughan D (2023). _dplyr: A Grammar of Data Manipulation_. https://dplyr.tidyverse.org, https://github.com/tidyverse/dplyr.
#> lavaan      |  0.6.13 |                                                                                                                                      Rosseel Y (2012). “lavaan: An R Package for Structural Equation Modeling.” _Journal of Statistical Software_, *48*(2), 1-36. doi:10.18637/jss.v048.i02 <https://doi.org/10.18637/jss.v048.i02>.
#> lme4        |  1.1.31 |                                                                                                                Bates D, Mächler M, Bolker B, Walker S (2015). “Fitting Linear Mixed-Effects Models Using lme4.” _Journal of Statistical Software_, *67*(1), 1-48. doi:10.18637/jss.v067.i01 <https://doi.org/10.18637/jss.v067.i01>.
#> performance |  0.10.2 |                                                         Lüdecke D, Ben-Shachar M, Patil I, Waggoner P, Makowski D (2021). “performance: An R Package for Assessment, Comparison and Testing of Statistical Models.” _Journal of Open Source Software_, *6*(60), 3139. doi:10.21105/joss.03139 <https://doi.org/10.21105/joss.03139>.
#> report      |   0.5.6 |                                                                                                          Makowski D, Lüdecke D, Patil I, Thériault R (2023). “Automated Results Reporting as a Practical Tool to Improve Reproducibility and Methodological Best Practices Adoption.” _CRAN_. <https://easystats.github.io/report/>.
summary(as.data.frame(r))
#> Package     | Version
#> ---------------------
#> Matrix      |   1.5.3
#> R           |   4.2.2
#> bayestestR  |  0.13.0
#> dplyr       |   1.1.0
#> lavaan      |  0.6.13
#> lme4        |  1.1.31
#> performance |  0.10.2
#> report      |   0.5.6

# Convenience functions
report_packages(include_R = FALSE)
#>   - Matrix (version 1.5.3; Bates D et al., 2022)
#>   - lme4 (version 1.1.31; Bates D et al., 2015)
#>   - performance (version 0.10.2; Lüdecke D et al., 2021)
#>   - bayestestR (version 0.13.0; Makowski D et al., 2019)
#>   - report (version 0.5.6; Makowski D et al., 2023)
#>   - lavaan (version 0.6.13; Rosseel Y, 2012)
#>   - dplyr (version 1.1.0; Wickham H et al., 2023)
cite_packages(prefix = "> ")
#> > Bates D, Maechler M, Jagan M (2022). _Matrix: Sparse and Dense Matrix Classes and Methods_. https://Matrix.R-forge.R-project.org/, https://Matrix.R-forge.R-project.org/doxygen/.
#> > Bates D, Mächler M, Bolker B, Walker S (2015). “Fitting Linear Mixed-Effects Models Using lme4.” _Journal of Statistical Software_, *67*(1), 1-48. doi:10.18637/jss.v067.i01 <https://doi.org/10.18637/jss.v067.i01>.
#> > Lüdecke D, Ben-Shachar M, Patil I, Waggoner P, Makowski D (2021). “performance: An R Package for Assessment, Comparison and Testing of Statistical Models.” _Journal of Open Source Software_, *6*(60), 3139. doi:10.21105/joss.03139 <https://doi.org/10.21105/joss.03139>.
#> > Makowski D, Ben-Shachar M, Lüdecke D (2019). “bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework.” _Journal of Open Source Software_, *4*(40), 1541. doi:10.21105/joss.01541 <https://doi.org/10.21105/joss.01541>, <https://joss.theoj.org/papers/10.21105/joss.01541>.
#> > Makowski D, Lüdecke D, Patil I, Thériault R (2023). “Automated Results Reporting as a Practical Tool to Improve Reproducibility and Methodological Best Practices Adoption.” _CRAN_. <https://easystats.github.io/report/>.
#> > R Core Team (2022). _R: A Language and Environment for Statistical Computing_. R Foundation for Statistical Computing, Vienna, Austria. <https://www.R-project.org/>.
#> > Rosseel Y (2012). “lavaan: An R Package for Structural Equation Modeling.” _Journal of Statistical Software_, *48*(2), 1-36. doi:10.18637/jss.v048.i02 <https://doi.org/10.18637/jss.v048.i02>.
#> > Wickham H, François R, Henry L, Müller K, Vaughan D (2023). _dplyr: A Grammar of Data Manipulation_. https://dplyr.tidyverse.org, https://github.com/tidyverse/dplyr.
report_system()
#> Analyses were conducted using the R Statistical language (version 4.2.2; R Core
#> Team, 2022) on Ubuntu 22.04.1 LTS