Reporting Bayesian Models

Create reports for Bayesian models. The description of the parameters follows the Sequential Effect eXistence and sIgnificance Testing framework (see SEXIT documentation).

Usage

# S3 method for class 'stanreg'
report(x, ...)

Arguments

x

Object of class lm or glm.

...

Arguments passed to or from other methods.

Value

An object of class report().

See also

Examples

# \donttest{
# Bayesian models
library(rstanarm)
#> This is rstanarm version 2.32.1
#> - See https://mc-stan.org/rstanarm/articles/priors for changes to default priors!
#> - Default priors may change, so it's safest to specify priors, even if equivalent to the defaults.
#> - For execution on a local, multicore CPU with excess RAM we recommend calling
#>   options(mc.cores = parallel::detectCores())
#> 
#> Attaching package: ‘rstanarm’
#> The following objects are masked from ‘package:brms’:
#> 
#>     dirichlet, exponential, get_y, lasso, ngrps
model <- suppressWarnings(stan_glm(mpg ~ qsec + wt, data = mtcars, refresh = 0, iter = 500))
r <- report(model)
r
#> We fitted a Bayesian linear model (estimated using MCMC sampling with 4 chains
#> of 500 iterations and a warmup of 250) to predict mpg with qsec and wt
#> (formula: mpg ~ qsec + wt). Priors over parameters were all set as normal (mean
#> = 0.00, SD = 8.43; mean = 0.00, SD = 15.40) distributions. The model's
#> explanatory power is substantial (R2 = 0.81, 95% CI [0.70, 0.89], adj. R2 =
#> 0.79). The model's intercept, corresponding to qsec = 0 and wt = 0, is at 19.96
#> (95% CI [8.68, 30.70]). Within this model:
#> 
#>   - The effect of qsec (Median = 0.92, 95% CI [0.39, 1.49]) has a 99.90%
#> probability of being positive (> 0), 99.20% of being significant (> 0.30), and
#> 0.00% of being large (> 1.81). The estimation successfully converged (Rhat =
#> 1.000) and the indices are reliable (ESS = 1189)
#>   - The effect of wt (Median = -5.04, 95% CI [-6.04, -3.98]) has a 100.00%
#> probability of being negative (< 0), 100.00% of being significant (< -0.30),
#> and 100.00% of being large (< -1.81). The estimation successfully converged
#> (Rhat = 0.999) and the indices are reliable (ESS = 1250)
#> 
#> Following the Sequential Effect eXistence and sIgnificance Testing (SEXIT)
#> framework, we report the median of the posterior distribution and its 95% CI
#> (Highest Density Interval), along the probability of direction (pd), the
#> probability of significance and the probability of being large. The thresholds
#> beyond which the effect is considered as significant (i.e., non-negligible) and
#> large are |0.30| and |1.81| (corresponding respectively to 0.05 and 0.30 of the
#> outcome's SD). Convergence and stability of the Bayesian sampling has been
#> assessed using R-hat, which should be below 1.01 (Vehtari et al., 2019), and
#> Effective Sample Size (ESS), which should be greater than 1000 (Burkner, 2017).
summary(r)
#> We fitted a Bayesian linear model to predict mpg with qsec and wt. Priors over
#> parameters were all set as normal (mean = 0.00, SD = 8.43; mean = 0.00, SD =
#> 15.40) distributions. The model's explanatory power is substantial (R2 = 0.81,
#> adj. R2 = 0.79). The model's intercept is at 19.96 (95% CI [8.68, 30.70]).
#> Within this model:
#> 
#>   - The effect of qsec (Median = 0.92, 95% CI [0.39, 1.49]) has 99.90%, 99.20%
#> and 0.00% probability of being positive (> 0), significant (> 0.30) and large
#> (> 1.81)
#>   - The effect of wt (Median = -5.04, 95% CI [-6.04, -3.98]) has 100.00%, 100.00%
#> and 100.00% probability of being negative (< 0), significant (< -0.30) and
#> large (< -1.81)
as.data.frame(r)
#> Parameter   | Median |         95% CI |     pd |  Rhat |     ESS |                   Prior |    Fit
#> ---------------------------------------------------------------------------------------------------
#> (Intercept) |  19.96 | [ 8.68, 30.70] |   100% | 1.000 | 1087.00 | Normal (20.09 +- 15.07) |       
#> qsec        |   0.92 | [ 0.39,  1.49] | 99.90% | 1.000 | 1189.00 |   Normal (0.00 +- 8.43) |       
#> wt          |  -5.04 | [-6.04, -3.98] |   100% | 0.999 | 1250.00 |  Normal (0.00 +- 15.40) |       
#>             |        |                |        |       |         |                         |       
#> ELPD        |        |                |        |       |         |                         | -79.13
#> LOOIC       |        |                |        |       |         |                         | 158.26
#> WAIC        |        |                |        |       |         |                         | 158.05
#> R2          |        |                |        |       |         |                         |   0.81
#> R2 (adj.)   |        |                |        |       |         |                         |   0.79
#> Sigma       |        |                |        |       |         |                         |   2.66
# }