Reporting Bayesian Models

Create reports for Bayesian models. The description of the parameters follows the Sequential Effect eXistence and sIgnificance Testing framework (see SEXIT documentation).

# S3 method for stanreg
report(x, ...)

Arguments

x
Object of class lm or glm.

...
Arguments passed to or from other methods.

Value

An object of class report().

See also

Examples

library(report)
# \donttest{
# Bayesian models
if (require("rstanarm")) {
  model <- stan_glm(mpg ~ qsec + wt, data = mtcars, refresh = 0, iter = 500)
  r <- report(model)
  r
  summary(r)
  as.data.frame(r)
  summary(as.data.frame(r))
}
#> Loading required package: rstanarm
#> Loading required package: Rcpp
#> This is rstanarm version 2.21.1
#> - See https://mc-stan.org/rstanarm/articles/priors for changes to default priors!
#> - Default priors may change, so it's safest to specify priors, even if equivalent to the defaults.
#> - For execution on a local, multicore CPU with excess RAM we recommend calling
#>   options(mc.cores = parallel::detectCores())
#> Parameter   | Median |         95% CI |     pd | % in ROPE |  Rhat |     ESS |                   Prior |  Fit
#> -------------------------------------------------------------------------------------------------------------
#> (Intercept) |  19.90 | [ 9.51, 30.27] |   100% |        0% | 0.998 |  922.00 | Normal (20.09 +- 15.07) |     
#> qsec        |   0.93 | [ 0.41,  1.47] | 99.90% |    10.52% | 0.998 | 1052.00 |   Normal (0.00 +- 8.43) |     
#> wt          |  -5.05 | [-6.10, -4.10] |   100% |        0% | 1.001 |  926.00 |  Normal (0.00 +- 15.40) |     
#>             |        |                |        |           |       |         |                         |     
#> R2          |        |                |        |           |       |         |                         | 0.81
#> R2 (adj.)   |        |                |        |           |       |         |                         | 0.78
#> Sigma       |        |                |        |           |       |         |                         | 2.67
# }