Reporting Bayesian Models

Create reports for Bayesian models. The description of the parameters follows the Sequential Effect eXistence and sIgnificance Testing framework (see SEXIT documentation).

# S3 method for stanreg
report(x, ...)

Arguments

x
Object of class lm or glm.

...
Arguments passed to or from other methods.

Value

A list-object of class report, which contains further list-objects with a short and long description of the model summary, as well as a short and long table of parameters and fit indices.

See also

Examples

library(report)

# Bayesian models
if (require("rstanarm")) {
  model <- stan_glm(mpg ~ qsec + wt, data = mtcars, refresh = 0, iter = 300, chains = 2, seed = 333)
  r <- report(model)
  r
  summary(r)
  as.data.frame(r)
  summary(as.data.frame(r))
}

#> Loading required package: rstanarm

#> Loading required package: Rcpp

#> This is rstanarm version 2.21.1

#> - See https://mc-stan.org/rstanarm/articles/priors for changes to default priors!

#> - Default priors may change, so it's safest to specify priors, even if equivalent to the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#>   options(mc.cores = parallel::detectCores())

#> 
#> Attaching package: ‘rstanarm’

#> The following object is masked from ‘package:performance’:
#> 
#>     pp_check

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: Note that the default rope range for linear models might change in future versions (see https://github.com/easystats/bayestestR/issues/364).Please set it explicitly to preserve current results.

#> Warning: Note that the default rope range for linear models might change in future versions (see https://github.com/easystats/bayestestR/issues/364).Please set it explicitly to preserve current results.

#> Parameter   | Median |         89% CI |   pd | % in ROPE |  Rhat |    ESS |                   Prior |  Fit
#> ----------------------------------------------------------------------------------------------------------
#> (Intercept) |  19.60 | [11.64, 30.05] | 100% |        0% | 0.997 | 151.00 | Normal (20.09 +- 15.07) |     
#> qsec        |   0.94 | [ 0.44,  1.36] | 100% |    13.67% | 0.996 | 173.00 |   Normal (0.00 +- 8.43) |     
#> wt          |  -4.98 | [-5.73, -4.25] | 100% |        0% | 0.998 | 156.00 |  Normal (0.00 +- 15.40) |     
#>             |        |                |      |           |       |        |                         |     
#> R2          |        |                |      |           |       |        |                         | 0.80
#> R2 (adj.)   |        |                |      |           |       |        |                         | 0.78
#> Sigma       |        |                |      |           |       |        |                         | 2.67