Bayesian Models Report

Create a report of Bayesian models.

# S3 method for stanreg
report(model, effsize = "default", ci = 0.89,
  standardize = "refit", standardize_robust = FALSE,
  centrality = "median", dispersion = FALSE, ci_method = "hdi",
  test = c("pd", "rope"), rope_range = "default", rope_ci = 1,
  bf_prior = NULL, diagnostic = c("ESS", "Rhat"),
  performance_metrics = "all", ...)

Arguments

model
Frequentist regression object.

effsize
Interpret the standardized parameters using a set of rules. Default corresponds to "funder2019" for linear models and "chen2010" for logistic models.

ci
Confidence Interval (CI) level. Default to 0.95 (95%).

standardize
Add standardized parameters. Can be FALSE or a character indicating the standardization method (see parameters_standardize).

standardize_robust
Robust standardization. See parameters_standardize.

centrality
The point-estimates (centrality indices) to compute. Character (vector) or list with one or more of these options: "median", "mean", "MAP" or "all".

dispersion
Logical, if TRUE, computes indices of dispersion related to the estimate(s) (SD and MAD for mean and median, respectively).

ci_method
The type of index used for Credible Interval. Can be "HDI" (default, see hdi) or "ETI" (see eti).

test
The indices of effect existence to compute. Character (vector) or list with one or more of these options: "p_direction" (or "pd"), "rope", "p_map", "equivalence_test" (or "equitest"), "bayesfactor" (or "bf") or "all" to compute all tests.#' For each "test", the corresponding bayestestR function is called (e.g. rope or p_direction) and its results included in the summary output.

rope_range
ROPE's lower and higher bounds. Should be a list of two values (e.g., c(-0.1, 0.1)) or "default". If "default", the bounds are set to x +- 0.1*SD(response).

rope_ci
The Credible Interval (CI) probability, corresponding to the proportion of HDI, to use for the percentage in ROPE.

bf_prior
Distribution representing a prior for the computation of Bayes factors. Used if the input is a posterior, otherwise (in the case of models) ignored.

diagnostic
Diagnostic metrics to compute. Character (vector) or list with one or more of these options: "ESS", "Rhat", "MCSE" or "all".

performance_metrics
See model_performance.

...
Arguments passed to or from other methods (e.g., to standardize).

Examples

library(report)
library(rstanarm)

#> Loading required package: Rcpp

#> Registered S3 method overwritten by 'xts':
#>   method     from
#>   as.zoo.xts zoo 

#> rstanarm (Version 2.18.2, packaged: 2018-11-08 22:19:38 UTC)

#> - Do not expect the default priors to remain the same in future rstanarm versions.

#> Thus, R scripts should specify priors explicitly, even if they are just the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#> options(mc.cores = parallel::detectCores())

#> - Plotting theme set to bayesplot::theme_default().

model <- rstanarm::stan_glm(Sepal.Length ~ Petal.Length * Species,
  data = iris, iter = 100, refresh = 0
)

#> Warning: The largest R-hat is 1.1, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: Markov chains did not converge! Do not analyze results!

r <- report(model)

#> Possible multicollinearity between Petal.Length:Speciesversicolor and Speciesversicolor (r = 0.87), Petal.Length:Speciesvirginica and Speciesvirginica (r = 0.75), Petal.Length:Speciesvirginica and Petal.Length:Speciesversicolor (r = 0.87). This might lead to inappropriate results. See 'Details' in '?rope'.

#> Warning: The largest R-hat is 1.36, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> Warning: Markov chains did not converge! Do not analyze results!

to_text(r)

#> We fitted a Bayesian linear model to predict Sepal.Length with Petal.Length and Species. The model's explanatory power is substantial (R2's median = 0.84, adj. R2 = 0.83). The model's intercept is at 3.93. Within this model:
#> 
#>   - The effect of Petal.Length has a probability of 99.50% of being positive and can be considered as very large and significant (median = 0.74, 89% CI [0.36, 1.05], 0% in ROPE, std. median = 1.01).
#>   - The effect of Species (versicolor) has a probability of 100% of being negative and can be considered as large and significant (median = -1.36, 89% CI [-2.37, -0.79], 0% in ROPE, std. median = -0.67).
#>   - The effect of Species (virginica) has a probability of 100% of being negative and can be considered as very large and significant (median = -2.70, 89% CI [-3.47, -1.95], 0% in ROPE, std. median = -1.62).
#>   - The interaction effect of Species (versicolor) on Petal.Length has a probability of 60.50% of being positive and can be considered as large and not significant (median = 0.09, 89% CI [-0.29, 0.46], 24.50% in ROPE, std. median = 0.72).
#>   - The interaction effect of Species (virginica) on Petal.Length has a probability of 88.50% of being positive and can be considered as very large and not significant (median = 0.24, 89% CI [-0.06, 0.61], 14.50% in ROPE, std. median = 1.19).

to_fulltext(r)

#> We fitted a Bayesian linear model (estimated using MCMC sampling with 4 chains of 100 iterations and a warmup of 50) to predict Sepal.Length with Petal.Length and Species (formula = Sepal.Length ~ Petal.Length * Species). Priors over parameters were set as normal (mean = 0.00, SD = 1.17), normal (mean = 0.00, SD = 2.07), normal (mean = 0.00, SD = 2.07), normal (mean = 0.00, SD = 1.02) and normal (mean = 0.00, SD = 0.78) distributions. The Region of Practical Equivalence (ROPE) percentage was defined as the proportion of the posterior distribution within the [-0.08, 0.08] range. The 89% Credible Intervals (CIs) were based on Highest Density Intervals (HDI). Standardized parameters were obtained by fitting the model on a standardized version of the dataset. Effect sizes were labelled following Funder's (2019) recommendations.
#> 
#> The model's explanatory power is substantial (R2's median = 0.84, 89% CI [0.82, 0.85], adj. R2 = 0.83). The model's intercept, corresponding to Sepal.Length = 0, Petal.Length = 0 and Species = setosa, is at 3.93 (89% CI [3.44, 4.49], 0% in ROPE, std. median = 0.30). Within this model:
#> 
#>   - The effect of Petal.Length has a probability of 99.50% of being positive and can be considered as very large and significant (median = 0.74, 89% CI [0.36, 1.05], 0% in ROPE, std. median = 1.01). However, the algorithm might not have successfuly converged (Rhat = 1.100) and the estimates cannot be considered as stable (ESS = 36).
#>   - The effect of Species (versicolor) has a probability of 100% of being negative and can be considered as large and significant (median = -1.36, 89% CI [-2.37, -0.79], 0% in ROPE, std. median = -0.67). However, the algorithm might not have successfuly converged (Rhat = 1.022) and the estimates cannot be considered as stable (ESS = 78).
#>   - The effect of Species (virginica) has a probability of 100% of being negative and can be considered as very large and significant (median = -2.70, 89% CI [-3.47, -1.95], 0% in ROPE, std. median = -1.62). However, the algorithm might not have successfuly converged (Rhat = 1.019) and the estimates cannot be considered as stable (ESS = 90).
#>   - The interaction effect of Species (versicolor) on Petal.Length has a probability of 60.50% of being positive and can be considered as large and not significant (median = 0.09, 89% CI [-0.29, 0.46], 24.50% in ROPE, std. median = 0.72). However, the algorithm might not have successfuly converged (Rhat = 1.075) and the estimates cannot be considered as stable (ESS = 48).
#>   - The interaction effect of Species (virginica) on Petal.Length has a probability of 88.50% of being positive and can be considered as very large and not significant (median = 0.24, 89% CI [-0.06, 0.61], 14.50% in ROPE, std. median = 1.19). However, the algorithm might not have successfuly converged (Rhat = 1.074) and the estimates cannot be considered as stable (ESS = 42).

to_table(r)

#> Parameter                      | Median |         89% CI |     pd | % in ROPE | Median (std.) |  Fit
#> ----------------------------------------------------------------------------------------------------
#> (Intercept)                    |   3.93 |   [3.44, 4.49] |   100% |        0% |          0.30 |     
#> Petal.Length                   |   0.74 |   [0.36, 1.05] | 99.50% |        0% |          1.01 |     
#> Speciesversicolor              |  -1.36 | [-2.37, -0.79] |   100% |        0% |         -0.67 |     
#> Speciesvirginica               |  -2.70 | [-3.47, -1.95] |   100% |        0% |         -1.62 |     
#> Petal.Length:Speciesversicolor |   0.09 |  [-0.29, 0.46] | 60.50% |    24.50% |          0.72 |     
#> Petal.Length:Speciesvirginica  |   0.24 |  [-0.06, 0.61] | 88.50% |    14.50% |          1.19 |     
#>                                |        |                |        |           |               |     
#> R2                             |        |                |        |           |               | 0.84
#> R2 (adj.)                      |        |                |        |           |               | 0.83

to_fulltable(r)

#> Parameter                           | Median |         89% CI |     pd | % in ROPE | ESS |  Rhat |              Prior | Median (std.) |    Fit
#> ----------------------------------------------------------------------------------------------------------------------------------------------
#> (Intercept)                         |   3.93 |   [3.44, 4.49] |   100% |        0% |  37 | 1.097 | Normal (0 +- 8.28) |          0.30 |       
#> Petal.Length                        |   0.74 |   [0.36, 1.05] | 99.50% |        0% |  36 | 1.100 | Normal (0 +- 1.17) |          1.01 |       
#> Species (versicolor)                |  -1.36 | [-2.37, -0.79] |   100% |        0% |  78 | 1.022 | Normal (0 +- 2.07) |         -0.67 |       
#> Species (virginica)                 |  -2.70 | [-3.47, -1.95] |   100% |        0% |  90 | 1.019 | Normal (0 +- 2.07) |         -1.62 |       
#> Petal.Length * Species (versicolor) |   0.09 |  [-0.29, 0.46] | 60.50% |    24.50% |  48 | 1.075 | Normal (0 +- 1.02) |          0.72 |       
#> Petal.Length * Species (virginica)  |   0.24 |  [-0.06, 0.61] | 88.50% |    14.50% |  42 | 1.074 | Normal (0 +- 0.78) |          1.19 |       
#>                                     |        |                |        |           |     |       |          NA ( +- ) |               |       
#> ELPD                                |        |                |        |           |     |       |          NA ( +- ) |               | -53.26
#> ELPD_SE                             |        |                |        |           |     |       |          NA ( +- ) |               |   8.20
#> LOOIC                               |        |                |        |           |     |       |          NA ( +- ) |               | 106.53
#> LOOIC_SE                            |        |                |        |           |     |       |          NA ( +- ) |               |  16.41
#> WAIC                                |        |                |        |           |     |       |          NA ( +- ) |               | 106.41
#> R2                                  |        |                |        |           |     |       |          NA ( +- ) |               |   0.84
#> R2 (adj.)                           |        |                |        |           |     |       |          NA ( +- ) |               |   0.83
#> RMSE                                |        |                |        |           |     |       |          NA ( +- ) |               |   0.35

model <- rstanarm::stan_lmer(Sepal.Length ~ Petal.Length + (1 | Species),
  data = iris, iter = 100, refresh = 0
)

#> Warning: The largest R-hat is 1.07, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

report(model)

#> Warning: The largest R-hat is 1.08, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#> We fitted a Bayesian linear mixed model to predict Sepal.Length with Petal.Length. The model included Species as random effects. The model's explanatory power is substantial (R2's median = 0.83, adj. R2 = 0.83). Within this model, the explanatory power related to the fixed effects alone (marginal R2's median) is of 0.74. The model's intercept is at 2.37. Within this model:
#> 
#>   - The effect of Petal.Length has a probability of 100% of being positive and can be considered as very large and significant (median = 0.88, 89% CI [0.76, 0.97], 0% in ROPE, std. median = 1.86).