Report the model type

Reports the type of different R objects (see list of supported objects in report).

report_model(x, table = NULL, ...)

Arguments

x
The R object that you want to report (see list of of supported objects above).

table
A table obtained via report_table(). If not provided, will run it.

...
Arguments passed to or from other methods.

Value

A character string.

Examples

library(report)

# h-tests
report_model(t.test(iris$Sepal.Width, iris$Sepal.Length))

#> Welch Two Sample t-test testing the difference between iris$Sepal.Width and iris$Sepal.Length (mean of x = 3.06, mean of y = 5.84)

report_model(cor.test(iris$Sepal.Width, iris$Sepal.Length))

#> Pearson's product-moment correlation between iris$Sepal.Width and iris$Sepal.Length

# ANOVA
report_model(aov(Sepal.Length ~ Species, data = iris))

#> For one-way between subjects designs, partial eta squared is equvilant to eta squared.
#> Returning eta squared.

#> ANOVA (formula: Sepal.Length ~ Species)

# GLMs
report_model(lm(Sepal.Length ~ Petal.Length * Species, data = iris))

#> linear model (estimated using OLS) to predict Sepal.Length with Petal.Length and Species (formula: Sepal.Length ~ Petal.Length * Species)

report_model(glm(vs ~ disp, data = mtcars, family = "binomial"))

#> logistic model (estimated using ML) to predict vs with disp (formula: vs ~ disp)

# Mixed models
if (require("lme4")) {
  model <- lme4::lmer(Sepal.Length ~ Petal.Length + (1 | Species), data = iris)
  report_model(model)
}

#> linear mixed model (estimated using REML and nloptwrap optimizer) to predict Sepal.Length with Petal.Length (formula: Sepal.Length ~ Petal.Length). The model included Species as random effect (formula: ~1 | Species)

# Bayesian models
if (require("rstanarm")) {
  model <- stan_glm(Sepal.Length ~ Species, data = iris, refresh = 0, iter = 600)
  report_model(model)
}

#> Warning: Note that the default rope range for linear models might change in future versions (see https://github.com/easystats/bayestestR/issues/364).Please set it explicitly to preserve current results.

#> Bayesian linear model (estimated using MCMC sampling with 4 chains of 600 iterations and a warmup of 300) to predict Sepal.Length with Species (formula: Sepal.Length ~ Species). Priors over parameters were all set as normal (mean = 0.00, SD = 4.38) distributions