Cite Loaded Packages

Citation table of loaded packages (show_packages includes version and name, and cite_packages includes only the citation).

show_packages(session)

cite_packages(session)

Arguments

session
A sessionInfo object.

Examples

show_packages(sessionInfo())

#>    Package Version
#> 3     lme4  1.1.21
#> 4   Matrix  1.2.17
#> 2     Rcpp   1.0.1
#> 5   report   0.1.0
#> 1 rstanarm  2.18.2
#> 6 testthat   2.1.1
#>                                                                                                                                                                                  References
#> 3 Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#> 4                        Douglas Bates and Martin Maechler (2019). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.2-17. https://CRAN.R-project.org/package=Matrix
#> 2                  Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18. URL http://www.jstatsoft.org/v40/i08/.
#> 5   Makowski, D. & Lüdecke, D. (2019). The report package for R: Ensuring the use of best practices for results reporting. CRAN. Available from https://github.com/easystats/report. doi: .
#> 1                                 Goodrich B, Gabry J, Ali I & Brilleman S. (2018). rstanarm: Bayesian applied regression modeling via Stan. R package version 2.17.4. http://mc-stan.org/.
#> 6                                                                                         Hadley Wickham. testthat: Get Started with Testing. The R Journal, vol. 3, no. 1, pp. 5--10, 2011

cite_packages(sessionInfo())

#>                                                                                                                                                              data[order(data$References), ]
#> 1                  Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18. URL http://www.jstatsoft.org/v40/i08/.
#> 2                        Douglas Bates and Martin Maechler (2019). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.2-17. https://CRAN.R-project.org/package=Matrix
#> 3 Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1-48. doi:10.18637/jss.v067.i01.
#> 4                                 Goodrich B, Gabry J, Ali I & Brilleman S. (2018). rstanarm: Bayesian applied regression modeling via Stan. R package version 2.17.4. http://mc-stan.org/.
#> 5                                                                                         Hadley Wickham. testthat: Get Started with Testing. The R Journal, vol. 3, no. 1, pp. 5--10, 2011
#> 6   Makowski, D. & Lüdecke, D. (2019). The report package for R: Ensuring the use of best practices for results reporting. CRAN. Available from https://github.com/easystats/report. doi: .